relm4

Module abstractions

A module for convenient abstractions over gtk-rs.

Re-exports

pub use drawing::DrawContext;

pub use drawing::DrawHandler;

Modules

drawing

Utility to help drawing on a gtk::DrawingArea in a Relm4 application. Create a DrawHandler, initialize it, and get its context when handling a message (that could be sent from the draw signal).

Structs

Toaster

An abstraction over adw::ToastOverlay that makes it easy to store it in the model of components.