relm4

Module binding

Easy data bindings between objects.

Particularly, this module provides types and traits that simplify connecting with the primary property of an object. In this context, primary means that the property is the most important and most accessed value stored by the object. Usually, this is very easy to figure out, such as the active property for gtk::ToggleButton. In any case, the primary property type are always documented.

To find out which widgets are currently supported and which property is considered as primary, please look at the list implementers for ConnectBinding. Contributions to add support for more widgets are always welcome.

Structs

BindingGuard

A RAII-guard that stores a value to a Binding.

BoolBinding

A data binding storing a value of type bool

F32Binding

A data binding storing a value of type f32

F64Binding

A data binding storing a value of type f64

I8Binding

A data binding storing a value of type i8

I32Binding

A data binding storing a value of type i32

I64Binding

A data binding storing a value of type i64

StringBinding

A data binding storing a value of type String

U8Binding

A data binding storing a value of type u8

U32Binding

A data binding storing a value of type u32

U64Binding

A data binding storing a value of type u64

Traits

Binding

A glib::Object with one primary property.

ConnectBinding

A trait that allows type-safe bindings between to the primary properties of two objects.

ConnectBindingExt

Extension for ConnectBinding. This trait is not implemented manually, but through automatically implemented for all types that implement ConnectBinding.