relm4::binding

Trait ConnectBinding

pub trait ConnectBinding {
    type Target;

    // Required method
    fn bind<B: Binding<Target = Self::Target>>(&self, binding: &B);
}

A trait that allows type-safe bindings between to the primary properties of two objects.

Required Associated Types

type Target

The type of the primary property.

Required Methods

fn bind<B: Binding<Target = Self::Target>>(&self, binding: &B)

Create a type-safe bidirectional between the primary property of an object and a Binding.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.