relm4::binding

Trait ConnectBindingExt

pub trait ConnectBindingExt {
    type Target;

    // Required method
    fn with_binding<B: Binding<Target = Self::Target>>(binding: &B) -> Self;
}

Extension for ConnectBinding. This trait is not implemented manually, but through automatically implemented for all types that implement ConnectBinding.

Required Associated Types

type Target

The type of the primary property. Inherited from ConnectBinding::Target.

Required Methods

fn with_binding<B: Binding<Target = Self::Target>>(binding: &B) -> Self

Create an object and immediately connect it with a Binding.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.