relm4

Module component

Components are smaller mostly independent parts of your application.

Modules

worker

A simpler version of components that does work in the background.

Structs

AsyncComponentBuilder

A component that is ready for docking and launch.

AsyncComponentParts

Contains the initial model and widgets being docked into a component.

AsyncComponentSender

Contains senders to send and receive messages from a Component.

AsyncComponentStream

Yields AsyncComponent::Output values as a stream and contains the input sender and the root widget.

AsyncConnector

Contains the post-launch input sender and output receivers with the root widget.

AsyncController

Controls the component from afar.

ComponentBuilder

A component that is ready for docking and launch.

ComponentParts

Contains the initial model and widgets being docked into a component.

ComponentStream

Yields Component::Output values as a stream and contains the input sender and the root widget.

Connector

Contains the post-launch input sender and output receivers with the root widget.

Controller

Controls the component from afar.

MessageBroker

A type that can be used in static variables to pass messages to components.

StateWatcher

Keeps track of a components model and view.

Traits

AsyncComponent

Asynchronous variant of Component.

AsyncComponentController

Shared behavior of component controller types.

Component

The fundamental building block of a Relm4 application.

ComponentController

Shared behavior of component controller types.

SimpleAsyncComponent

Asynchronous variant of SimpleComponent.

SimpleComponent

Elm-style variant of a Component with view updates separated from input updates.

Type Aliases

CommandFuture

A future returned by a component’s command method.