relm4::component

Struct AsyncComponentBuilder

pub struct AsyncComponentBuilder<C: AsyncComponent> {
    pub root: C::Root,
    /* private fields */
}

A component that is ready for docking and launch.

Fields
root: C::Root

The root widget of the component.

Implementations

impl<C: AsyncComponent> AsyncComponentBuilder<C>

pub fn update_root<F: FnOnce(&mut C::Root)>(self, func: F) -> Self

Configure the root widget before launching.

pub const fn widget(&self) -> &C::Root

Access the root widget before the component is initialized.

pub fn priority(self, priority: Priority) -> Self

Change the priority at which the messages of this component are handled.

Use glib::Priority::HIGH for high priority event sources.

Use glib::Priority::LOW for very low priority background tasks.

Use glib::Priority::DEFAULT_IDLE for default priority idle functions.

Use glib::Priority::HIGH_IDLE for high priority idle functions.

impl<C: AsyncComponent> AsyncComponentBuilder<C>
where C::Root: AsRef<Widget>,

pub fn attach_to<T: RelmContainerExt<Child = Widget>>( self, container: &T, ) -> Self

Attach the component’s root widget to a given container.

impl<C: AsyncComponent> AsyncComponentBuilder<C>
where C::Root: AsRef<Window> + Clone,

pub fn transient_for(self, widget: impl AsRef<Widget>) -> Self

Set the component’s root widget transient for a given window. This function doesn’t require a gtk::Window as parameter, but instead uses RelmWidgetExt::toplevel_window() to retrieve the toplevel window of any gtk::Widget. Therefore, you don’t have to pass a window to every component.

If the root widget is a native dialog, such as gtk::FileChooserNative, you should use transient_for_native instead.

impl<C: AsyncComponent> AsyncComponentBuilder<C>
where C::Root: AsRef<NativeDialog> + Clone,

pub fn transient_for_native(self, widget: impl AsRef<Widget>) -> Self

Set the component’s root widget transient for a given window. This function doesn’t require a gtk::Window as parameter, but instead uses RelmWidgetExt::toplevel_window() to retrieve the toplevel window of any gtk::Widget. Therefore, you don’t have to pass a window to every component.

Applicable to native dialogs only, such as gtk::FileChooserNative. If the root widget is a non-native dialog, you should use transient_for instead.

impl<C: AsyncComponent> AsyncComponentBuilder<C>

pub fn launch(self, payload: C::Init) -> AsyncConnector<C>

Starts the component, passing ownership to a future attached to a gtk::glib::MainContext.

pub fn launch_with_broker( self, payload: C::Init, broker: &MessageBroker<C::Input>, ) -> AsyncConnector<C>

Similar to launch() but also initializes a MessageBroker.

Panics

This method panics if the message broker was already initialized in another launch.