Struct relm4::component::MessageBroker

pub struct MessageBroker<C: Component> { /* private fields */ }

A type that can be used in static variables to pass messages to components.

The primary use-case for this type is to communicate between components on different levels.

Imagine you have three components: A, B and C. A and B are children of the main application, but C is a child of B. If C wants to pass a message to A, it relies on B to forward that message to A. This is not great because B has nothing to do this message but has to implement additional logic only to pass the message through. MessageBroker allows you to use statics to remove this limitation.

Note

MessageBroker will not forward any messages until you initialize them with ComponentBuilder::launch_with_broker().

Only initialize the message broker once!

use relm4::{MessageBroker, Component};

static MY_COMPONENT: MessageBroker<MyComponent> = MessageBroker::new();

// Initialize the component and the message broker with `launch_with_broker`.
let controller = MyComponent::builder().launch_with_broker((), &MY_COMPONENT).detach();

Implementations

impl<C: Component> MessageBroker<C>

pub const fn new() -> Self

Creates a new MessageBroker.

The returned message broker will not forward messages until it’s initialized.

Examples found in repository?

relm4/examples/message_broker.rs (line 10)

10

static HEADER_BROKER: MessageBroker<Header> = MessageBroker::new();

Hide additional examples

relm4/examples/transient_dialog.rs (line 10)

10

static DIALOG_BROKER: MessageBroker<Dialog> = MessageBroker::new();

pub fn sender(&self) -> &Sender<C::Input>

Get the input sender of the component.

pub fn send(&self, input: C::Input)

Send an input message to the component.