relm4::component::worker

Trait Worker

pub trait Worker:
    Sized
    + Send
    + 'static {
    type Init: 'static + Send;
    type Input: 'static + Send + Debug;
    type Output: 'static + Send + Debug;

    // Required methods
    fn init(init: Self::Init, sender: ComponentSender<Self>) -> Self;
    fn update(&mut self, message: Self::Input, sender: ComponentSender<Self>);
}

Receives inputs and outputs in the background.

All types that implement Worker will also implement Component automatically.

If you need more flexibility when using workers, you can simply implement Component instead and set the Component::Widgets and Component::Root types both to (). This will still allow you to use all worker related methods because internally a worker is just seen as a Component without widgets.

Required Associated Types

type Init: 'static + Send

The initial parameters that will be used to build the worker state.

type Input: 'static + Send + Debug

The type of inputs that this worker shall receive.

type Output: 'static + Send + Debug

The type of outputs that this worker shall send.

Required Methods

fn init(init: Self::Init, sender: ComponentSender<Self>) -> Self

Defines the initial state of the worker.

fn update(&mut self, message: Self::Input, sender: ComponentSender<Self>)

Defines how inputs will bep processed

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.