Module relm4::drawing

Utility to help drawing on a gtk::DrawingArea in a Relm4 application. Create a DrawHandler, initialize it, and get its context when handling a message (that could be sent from the draw signal).

Structs

DrawContext

Context returned by DrawHandler that stores a Context along with additional data required for drawing.

DrawHandler

Manager for drawing operations.