relm4

Module factory

Defines traits and data types to generate widgets from collections efficiently.

Re-exports

pub use widgets::traits::*;

Modules

positions

Position types for various widgets.

widgets

Traits and implementations used for factories to interact with widgets.

Structs

AsyncFactorySender

Contains senders to send and receive messages from a Component.

AsyncFactoryVecDeque

A container similar to VecDeque that can be used to store data associated with components that implement AsyncFactoryComponent.

AsyncFactoryVecDequeBuilder

A builder-pattern struct for building a AsyncFactoryVecDeque.

AsyncFactoryVecDequeConnector

Second stage of the builder-pattern for building a AsyncFactoryVecDeque.

AsyncFactoryVecDequeGuard

Provides methods to edit the underlying AsyncFactoryVecDeque.

DynamicIndex

A dynamic index that updates automatically when items are shifted inside a factory container.

FactoryHashMap

A container similar to HashMap that can be used to store values of type FactoryComponent.

FactoryHashMapBuilder

A builder-pattern struct for building a FactoryHashMap.

FactoryHashMapConnector

Second stage of the builder-pattern for building a FactoryHashMap.

FactorySender

Contains senders to send and receive messages from a Component.

FactoryVecDeque

A container similar to VecDeque that can be used to store data associated with components that implement FactoryComponent.

FactoryVecDequeBuilder

A builder-pattern struct for building a FactoryVecDeque.

FactoryVecDequeConnector

Second stage of the builder-pattern for building a FactoryVecDeque.

FactoryVecDequeGuard

Provides methods to edit the underlying FactoryVecDeque.

Traits

AsyncFactoryComponent

A component that’s stored inside a factory. Similar to Component but adjusted to fit the life cycle of factories.

CloneableFactoryComponent

Extension for FactoryComponent that makes elements cloneable.

FactoryComponent

A component that’s stored inside a factory. Similar to Component but adjusted to fit the life cycle of factories.