Struct relm4::factory::collections::FactoryVecDeque

pub struct FactoryVecDeque<Widget, C, ParentMsg> where
    Widget: FactoryView,
    C: FactoryComponent<Widget, ParentMsg> + Position<Widget::Position>,
    C::Root: AsRef<Widget::Children>,
    ParentMsg: 'static,
    Widget: 'static,  { /* private fields */ }

A container similar to VecDeque that can be used to store data associated with components that implement FactoryComponent.

Implementations

impl<Widget, C, ParentMsg> FactoryVecDeque<Widget, C, ParentMsg> where
Widget: FactoryView,
C: FactoryComponent<Widget, ParentMsg> + Position<Widget::Position>,
C::Root: AsRef<Widget::Children>,

pub fn new(widget: Widget, parent_sender: &Sender<ParentMsg>) -> Self

Creates a new FactoryVecDeque.

Examples found in repository?

examples/libadwaita/tab_factory.rs (line 238)

↕

228
229
230
231
232
233
234
235
236
237
238
239
240
241
242

    fn init(
        counter: Self::InitParams,
        root: &Self::Root,
        sender: &ComponentSender<Self>,
    ) -> ComponentParts<Self> {
        // Insert the macro codegen here
        let widgets = view_output!();

        let model = AppModel {
            created_widgets: counter,
            counters: FactoryVecDeque::new(widgets.tabs.clone(), &sender.input),
        };

        ComponentParts { model, widgets }
    }

Hide additional examples

examples/factory.rs (line 209)

↕

199
200
201
202
203
204
205
206
207
208
209
210
211
212
213

    fn init(
        counter: Self::InitParams,
        root: &Self::Root,
        sender: &ComponentSender<Self>,
    ) -> ComponentParts<Self> {
        // Insert the macro codegen here
        let widgets = view_output!();

        let model = AppModel {
            created_widgets: counter,
            counters: FactoryVecDeque::new(widgets.counter_box.clone(), &sender.input),
        };

        ComponentParts { model, widgets }
    }

examples/grid_factory.rs (line 246)

↕

236
237
238
239
240
241
242
243
244
245
246
247
248
249
250

    fn init(
        counter: Self::InitParams,
        root: &Self::Root,
        sender: &ComponentSender<Self>,
    ) -> ComponentParts<Self> {
        // Insert the macro codegen here
        let widgets = view_output!();

        let model = AppModel {
            created_widgets: counter,
            counters: FactoryVecDeque::new(widgets.counter_box.clone(), &sender.input),
        };

        ComponentParts { model, widgets }
    }

pub fn render_changes(&self)

Updates the widgets according to the changes made to the factory. All updates accumulate until this method is called and are handled efficiently.

For example, swapping two elements twice will only swap the data twice, but won’t cause any UI updates.

Also, only modified elements will be updated.

Examples found in repository?

examples/grid_factory.rs (line 280)

↕

252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }
        self.counters.render_changes();
    }

Hide additional examples

examples/factory.rs (line 244)

↕

215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }

        self.counters.render_changes();
    }

examples/libadwaita/tab_factory.rs (line 279)

↕

244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        self.counters.apply_external_updates();
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }
        if self.counters.len() > 2 {
            // Testing different stuff...
            //self.counters.swap(0, 2);
            // let mut counter = self.counters.get_mut(1);
            // counter.value = counter.value.wrapping_add(10);
        }
        self.counters.render_changes();
    }

pub fn len(&self) -> usize

Returns the number of elements in the FactoryVecDeque.

Examples found in repository?

examples/grid_factory.rs (line 268)

↕

252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }
        self.counters.render_changes();
    }

Hide additional examples

examples/factory.rs (line 231)

↕

215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }

        self.counters.render_changes();
    }

examples/libadwaita/tab_factory.rs (line 261)

≺ ≻↕

244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        self.counters.apply_external_updates();
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }
        if self.counters.len() > 2 {
            // Testing different stuff...
            //self.counters.swap(0, 2);
            // let mut counter = self.counters.get_mut(1);
            // counter.value = counter.value.wrapping_add(10);
        }
        self.counters.render_changes();
    }

pub fn is_empty(&self) -> bool

Returns true if the FactoryVecDeque is empty.

pub fn send(&self, index: usize, msg: C::Input)

Send a message to one of the elements.

pub fn try_get(&self, index: usize) -> Option<Ref<'_, C>>

Tries to get an immutable reference to the model of one element.

Returns None is index is invalid.

Panics

Panics when the same element was borrowed mutably somewhere else.

pub fn try_get_mut(&mut self, index: usize) -> Option<RefMut<'_, C>>

Tries to get a mutable reference to the model of one element.

Returns None is index is invalid.

Panics

Panics when the same element was borrowed somewhere else.

pub fn get(&self, index: usize) -> Ref<'_, C>

Provides a reference to the model of one element.

Element at index 0 is the front of the queue.

Panics

Panics when the index is invalid or the same element was borrowed mutably somewhere else.

pub fn get_mut(&mut self, index: usize) -> RefMut<'_, C>

Provides a mutable reference to the model of one element.

Element at index 0 is the front of the queue.

Panics

Panics when the index is invalid or the same element was borrowed somewhere else.

pub fn back(&self) -> Option<Ref<'_, C>>

Provides a reference to the model of the back element.

Returns None if the deque is empty.

pub fn front(&self) -> Option<Ref<'_, C>>

Provides a reference to the model of the front element.

Returns None if the deque is empty.

pub fn back_mut(&mut self) -> Option<RefMut<'_, C>>

Provides a mutable reference to the model of the back element.

Returns None if the deque is empty.

pub fn front_mut(&mut self) -> Option<RefMut<'_, C>>

Provides a mutable reference to the model of the front element.

Returns None if the deque is empty.

pub fn pop_back(&mut self) -> Option<C>

Removes the last element from the FactoryVecDeque and returns it, or None if it is empty.

Examples found in repository?

examples/grid_factory.rs (line 259)

↕

252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }
        self.counters.render_changes();
    }

Hide additional examples

examples/factory.rs (line 222)

↕

215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }

        self.counters.render_changes();
    }

examples/libadwaita/tab_factory.rs (line 252)

↕

244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        self.counters.apply_external_updates();
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }
        if self.counters.len() > 2 {
            // Testing different stuff...
            //self.counters.swap(0, 2);
            // let mut counter = self.counters.get_mut(1);
            // counter.value = counter.value.wrapping_add(10);
        }
        self.counters.render_changes();
    }

pub fn pop_front(&mut self) -> Option<C>

Removes the first element from the FactoryVecDeque and returns it, or None if it is empty.

pub fn remove(&mut self, index: usize) -> Option<C>

Removes and returns the element at index from the FactoryVecDeque. Returns None if index is out of bounds.

Element at index 0 is the front of the queue.

pub fn widget(&self) -> &Widget

Returns the widget all components are attached to.

pub fn push_back(&mut self, init_params: C::InitParams)

Appends an element at the end of the FactoryVecDeque.

Examples found in repository?

examples/grid_factory.rs (line 255)

↕

252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }
        self.counters.render_changes();
    }

Hide additional examples

examples/factory.rs (line 218)

↕

215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }

        self.counters.render_changes();
    }

examples/libadwaita/tab_factory.rs (line 248)

↕

244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        self.counters.apply_external_updates();
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }
        if self.counters.len() > 2 {
            // Testing different stuff...
            //self.counters.swap(0, 2);
            // let mut counter = self.counters.get_mut(1);
            // counter.value = counter.value.wrapping_add(10);
        }
        self.counters.render_changes();
    }

pub fn push_front(&mut self, init_params: C::InitParams)

Prepends an element to the FactoryVecDeque.

pub fn insert(&mut self, index: usize, init_params: C::InitParams)

Inserts an element at index within the FactoryVecDeque, shifting all elements with indices greater than or equal to index towards the back.

Element at index 0 is the front of the queue.

Panics

Panics if index is greater than FactoryVecDeque’s length.

pub fn swap(&mut self, first: usize, second: usize)

Swaps elements at indices first and second.

first and second may be equal.

Element at index 0 is the front of the queue.

Panics

Panics if either index is out of bounds.

pub fn move_to(&mut self, current_position: usize, target: usize)

Moves an element at index current_position to target, shifting all elements between these positions.

current_position and target may be equal.

Element at index 0 is the front of the queue.

Panics

Panics if either index is out of bounds.

Examples found in repository?

examples/grid_factory.rs (line 269)

≺ ≻↕

252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }
        self.counters.render_changes();
    }

Hide additional examples

examples/factory.rs (line 232)

≺ ≻↕

215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }

        self.counters.render_changes();
    }

examples/libadwaita/tab_factory.rs (line 262)

≺ ≻↕

244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        self.counters.apply_external_updates();
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }
        if self.counters.len() > 2 {
            // Testing different stuff...
            //self.counters.swap(0, 2);
            // let mut counter = self.counters.get_mut(1);
            // counter.value = counter.value.wrapping_add(10);
        }
        self.counters.render_changes();
    }

pub fn move_front(&mut self, current_position: usize)

Moves an element at index current_position to the front, shifting all elements between these positions.

Panics

Panics if index is out of bounds.

Examples found in repository?

examples/grid_factory.rs (line 262)

↕

252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }
        self.counters.render_changes();
    }

Hide additional examples

examples/factory.rs (line 225)

↕

215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }

        self.counters.render_changes();
    }

examples/libadwaita/tab_factory.rs (line 255)

↕

244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        self.counters.apply_external_updates();
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }
        if self.counters.len() > 2 {
            // Testing different stuff...
            //self.counters.swap(0, 2);
            // let mut counter = self.counters.get_mut(1);
            // counter.value = counter.value.wrapping_add(10);
        }
        self.counters.render_changes();
    }

pub fn move_back(&mut self, current_position: usize)

Moves an element at index current_position to the back, shifting all elements between these positions.

Panics

Panics if index is out of bounds.

pub fn clear(&mut self)

Remove all components from the FactoryVecDeque.

impl<C, ParentMsg> FactoryVecDeque<TabView, C, ParentMsg> where
C: FactoryComponent<TabView, ParentMsg> + Position<()>,
C::Root: AsRef<Widget>,

pub fn apply_external_updates(&mut self)

Apply external updates that happened between the last render.

**YOU MUST NOT EDIT THE FactoryVecDeque BETWEEN CALLING Self::render_changes AND THIS METHOD. THIS MIGHT CAUSE UNDEFINED BEHAVIOR.

Examples found in repository?

examples/libadwaita/tab_factory.rs (line 245)

↕

244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280

    fn update(&mut self, msg: Self::Input, _sender: &ComponentSender<Self>) {
        self.counters.apply_external_updates();
        match msg {
            AppMsg::AddCounter => {
                self.counters.push_back(self.created_widgets);
                self.created_widgets = self.created_widgets.wrapping_add(1);
            }
            AppMsg::RemoveCounter => {
                self.counters.pop_back();
            }
            AppMsg::SendFront(index) => {
                self.counters.move_front(index.current_index());
            }
            AppMsg::MoveDown(index) => {
                let index = index.current_index();
                let new_index = index + 1;
                // Already at the end?
                if new_index < self.counters.len() {
                    self.counters.move_to(index, new_index);
                }
            }
            AppMsg::MoveUp(index) => {
                let index = index.current_index();
                // Already at the start?
                if index != 0 {
                    self.counters.move_to(index, index - 1);
                }
            }
        }
        if self.counters.len() > 2 {
            // Testing different stuff...
            //self.counters.swap(0, 2);
            // let mut counter = self.counters.get_mut(1);
            // counter.value = counter.value.wrapping_add(10);
        }
        self.counters.render_changes();
    }