relm4::factory

Struct AsyncFactoryVecDequeGuard

pub struct AsyncFactoryVecDequeGuard<'a, C: AsyncFactoryComponent>
where
    <C::ParentWidget as FactoryView>::ReturnedWidget: Clone,
{ /* private fields */ }

Provides methods to edit the underlying AsyncFactoryVecDeque.

The changes will be rendered on the widgets after the guard goes out of scope.

Implementations

impl<'a, C: AsyncFactoryComponent> AsyncFactoryVecDequeGuard<'a, C>
where <C::ParentWidget as FactoryView>::ReturnedWidget: Clone,

pub fn drop(self)

Drops the guard and renders all changes.

Use this to transfer full ownership back to the AsyncFactoryVecDeque.

pub fn get_mut(&mut self, index: usize) -> Option<&mut C>

Tries to get a mutable reference to the model of one element.

Returns None if index is invalid or the async init_model() method hasn’t returned yet.

pub fn back_mut(&mut self) -> Option<&mut C>

Provides a mutable reference to the model of the back element.

Returns None if the deque is empty or the async init_model() method of the last element hasn’t returned yet.

pub fn front_mut(&mut self) -> Option<&mut C>

Provides a mutable reference to the model of the front element.

Returns None if the deque is empty or the async init_model() method of the first element hasn’t returned yet.

pub fn pop_back(&mut self) -> Option<C>

Removes the last element from the AsyncFactoryVecDeque and returns it, or None if it is empty or the async init_model() method of the element hasn’t returned yet.

Examples found in repository?

relm4/examples/factory_async.rs (line 221)

213    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
214        let mut counters_guard = self.counters.guard();
215        match msg {
216            AppMsg::AddCounter => {
217                counters_guard.push_back(self.created_widgets);
218                self.created_widgets = self.created_widgets.wrapping_add(1);
219            }
220            AppMsg::RemoveCounter => {
221                counters_guard.pop_back();
222            }
223            AppMsg::SendFront(index) => {
224                counters_guard.move_front(index.current_index());
225            }
226            AppMsg::MoveDown(index) => {
227                let index = index.current_index();
228                let new_index = index + 1;
229                // Already at the end?
230                if new_index < counters_guard.len() {
231                    counters_guard.move_to(index, new_index);
232                }
233            }
234            AppMsg::MoveUp(index) => {
235                let index = index.current_index();
236                // Already at the start?
237                if index != 0 {
238                    counters_guard.move_to(index, index - 1);
239                }
240            }
241            AppMsg::Remove(index) => {
242                counters_guard.remove(index.current_index());
243            }
244        }
245    }

pub fn pop_front(&mut self) -> Option<C>

Removes the first element from the AsyncFactoryVecDeque and returns it, or None if it is empty or the async init_model() method of the element hasn’t returned yet.

pub fn remove(&mut self, index: usize) -> Option<C>

Removes and returns the element at index from the AsyncFactoryVecDeque. or None if it is empty or the async init_model() method of the element hasn’t returned yet.

Element at index 0 is the front of the queue.

Examples found in repository?

relm4/examples/factory_async.rs (line 242)

213    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
214        let mut counters_guard = self.counters.guard();
215        match msg {
216            AppMsg::AddCounter => {
217                counters_guard.push_back(self.created_widgets);
218                self.created_widgets = self.created_widgets.wrapping_add(1);
219            }
220            AppMsg::RemoveCounter => {
221                counters_guard.pop_back();
222            }
223            AppMsg::SendFront(index) => {
224                counters_guard.move_front(index.current_index());
225            }
226            AppMsg::MoveDown(index) => {
227                let index = index.current_index();
228                let new_index = index + 1;
229                // Already at the end?
230                if new_index < counters_guard.len() {
231                    counters_guard.move_to(index, new_index);
232                }
233            }
234            AppMsg::MoveUp(index) => {
235                let index = index.current_index();
236                // Already at the start?
237                if index != 0 {
238                    counters_guard.move_to(index, index - 1);
239                }
240            }
241            AppMsg::Remove(index) => {
242                counters_guard.remove(index.current_index());
243            }
244        }
245    }

pub fn push_back(&mut self, init: C::Init) -> DynamicIndex

Appends an element at the end of the AsyncFactoryVecDeque.

Examples found in repository?

relm4/examples/factory_async.rs (line 217)

213    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
214        let mut counters_guard = self.counters.guard();
215        match msg {
216            AppMsg::AddCounter => {
217                counters_guard.push_back(self.created_widgets);
218                self.created_widgets = self.created_widgets.wrapping_add(1);
219            }
220            AppMsg::RemoveCounter => {
221                counters_guard.pop_back();
222            }
223            AppMsg::SendFront(index) => {
224                counters_guard.move_front(index.current_index());
225            }
226            AppMsg::MoveDown(index) => {
227                let index = index.current_index();
228                let new_index = index + 1;
229                // Already at the end?
230                if new_index < counters_guard.len() {
231                    counters_guard.move_to(index, new_index);
232                }
233            }
234            AppMsg::MoveUp(index) => {
235                let index = index.current_index();
236                // Already at the start?
237                if index != 0 {
238                    counters_guard.move_to(index, index - 1);
239                }
240            }
241            AppMsg::Remove(index) => {
242                counters_guard.remove(index.current_index());
243            }
244        }
245    }

pub fn push_front(&mut self, init: C::Init) -> DynamicIndex

Prepends an element to the AsyncFactoryVecDeque.

pub fn insert(&mut self, index: usize, init: C::Init) -> DynamicIndex

Inserts an element at index within the AsyncFactoryVecDeque, shifting all elements with indices greater than or equal to index towards the back.

Element at index 0 is the front of the queue.

Panics

Panics if index is greater than AsyncFactoryVecDeque’s length.

pub fn swap(&mut self, first: usize, second: usize)

Swaps elements at indices first and second.

first and second may be equal.

Element at index 0 is the front of the queue.

Panics

Panics if either index is out of bounds.

pub fn move_to(&mut self, current_position: usize, target: usize)

Moves an element at index current_position to target, shifting all elements between these positions.

current_position and target may be equal.

Element at index 0 is the front of the queue.

Panics

Panics if either index is out of bounds.

Examples found in repository?

relm4/examples/factory_async.rs (line 231)

213    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
214        let mut counters_guard = self.counters.guard();
215        match msg {
216            AppMsg::AddCounter => {
217                counters_guard.push_back(self.created_widgets);
218                self.created_widgets = self.created_widgets.wrapping_add(1);
219            }
220            AppMsg::RemoveCounter => {
221                counters_guard.pop_back();
222            }
223            AppMsg::SendFront(index) => {
224                counters_guard.move_front(index.current_index());
225            }
226            AppMsg::MoveDown(index) => {
227                let index = index.current_index();
228                let new_index = index + 1;
229                // Already at the end?
230                if new_index < counters_guard.len() {
231                    counters_guard.move_to(index, new_index);
232                }
233            }
234            AppMsg::MoveUp(index) => {
235                let index = index.current_index();
236                // Already at the start?
237                if index != 0 {
238                    counters_guard.move_to(index, index - 1);
239                }
240            }
241            AppMsg::Remove(index) => {
242                counters_guard.remove(index.current_index());
243            }
244        }
245    }

pub fn move_front(&mut self, current_position: usize)

Moves an element at index current_position to the front, shifting all elements between these positions.

Panics

Panics if index is out of bounds.

Examples found in repository?

relm4/examples/factory_async.rs (line 224)

213    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
214        let mut counters_guard = self.counters.guard();
215        match msg {
216            AppMsg::AddCounter => {
217                counters_guard.push_back(self.created_widgets);
218                self.created_widgets = self.created_widgets.wrapping_add(1);
219            }
220            AppMsg::RemoveCounter => {
221                counters_guard.pop_back();
222            }
223            AppMsg::SendFront(index) => {
224                counters_guard.move_front(index.current_index());
225            }
226            AppMsg::MoveDown(index) => {
227                let index = index.current_index();
228                let new_index = index + 1;
229                // Already at the end?
230                if new_index < counters_guard.len() {
231                    counters_guard.move_to(index, new_index);
232                }
233            }
234            AppMsg::MoveUp(index) => {
235                let index = index.current_index();
236                // Already at the start?
237                if index != 0 {
238                    counters_guard.move_to(index, index - 1);
239                }
240            }
241            AppMsg::Remove(index) => {
242                counters_guard.remove(index.current_index());
243            }
244        }
245    }

pub fn move_back(&mut self, current_position: usize)

Moves an element at index current_position to the back, shifting all elements between these positions.

Panics

Panics if index is out of bounds.

pub fn clear(&mut self)

Remove all components from the AsyncFactoryVecDeque.

pub fn iter_mut( &mut self, ) -> impl DoubleEndedIterator<Item = Option<&mut C>> + ExactSizeIterator + FusedIterator

Returns an iterator over the components that returns mutable references.

Each item will be Some if the async init_model() method of the item returned and otherwise None.