relm4::factory

Trait AsyncFactoryComponent

pub trait AsyncFactoryComponent:
    AsyncPosition<<Self::ParentWidget as FactoryView>::Position>
    + Sized
    + 'static {
    type ParentWidget: FactoryView + 'static;
    type CommandOutput: Debug + Send + 'static;
    type Input: Debug + 'static;
    type Output: Debug + 'static;
    type Init;
    type Root: AsRef<<Self::ParentWidget as FactoryView>::Children> + Debug + Clone;
    type Widgets: 'static;

    // Required methods
    fn init_model(
        init: Self::Init,
        index: &DynamicIndex,
        sender: AsyncFactorySender<Self>,
    ) -> impl Future<Output = Self>;
    fn init_root() -> Self::Root;
    fn init_widgets(
        &mut self,
        index: &DynamicIndex,
        root: Self::Root,
        returned_widget: &<Self::ParentWidget as FactoryView>::ReturnedWidget,
        sender: AsyncFactorySender<Self>,
    ) -> Self::Widgets;

    // Provided methods
    fn init_loading_widgets(_root: Self::Root) -> Option<LoadingWidgets> { ... }
    fn update(
        &mut self,
        message: Self::Input,
        sender: AsyncFactorySender<Self>,
    ) -> impl Future<Output = ()> { ... }
    fn update_cmd(
        &mut self,
        message: Self::CommandOutput,
        sender: AsyncFactorySender<Self>,
    ) -> impl Future<Output = ()> { ... }
    fn update_cmd_with_view(
        &mut self,
        widgets: &mut Self::Widgets,
        message: Self::CommandOutput,
        sender: AsyncFactorySender<Self>,
    ) -> impl Future<Output = ()> { ... }
    fn update_view(
        &self,
        widgets: &mut Self::Widgets,
        sender: AsyncFactorySender<Self>,
    ) { ... }
    fn update_with_view(
        &mut self,
        widgets: &mut Self::Widgets,
        message: Self::Input,
        sender: AsyncFactorySender<Self>,
    ) -> impl Future<Output = ()> { ... }
    fn shutdown(
        &mut self,
        widgets: &mut Self::Widgets,
        output: Sender<Self::Output>,
    ) { ... }
    fn id(&self) -> String { ... }
}

A component that’s stored inside a factory. Similar to Component but adjusted to fit the life cycle of factories.

Required Associated Types

type ParentWidget: FactoryView + 'static

Container widget to which all widgets of the factory will be added.

type CommandOutput: Debug + Send + 'static

Messages which are received from commands executing in the background.

type Input: Debug + 'static

The message type that the factory component accepts as inputs.

type Output: Debug + 'static

The message type that the factory component provides as outputs.

type Init

The parameter used to initialize the factory component.

type Root: AsRef<<Self::ParentWidget as FactoryView>::Children> + Debug + Clone

The top-level widget of the factory component.

type Widgets: 'static

The type that’s used for storing widgets created for this factory component.

Required Methods

fn init_model( init: Self::Init, index: &DynamicIndex, sender: AsyncFactorySender<Self>, ) -> impl Future<Output = Self>

Initializes the model.

fn init_root() -> Self::Root

Initializes the root widget

fn init_widgets( &mut self, index: &DynamicIndex, root: Self::Root, returned_widget: &<Self::ParentWidget as FactoryView>::ReturnedWidget, sender: AsyncFactorySender<Self>, ) -> Self::Widgets

Initializes the widgets.

Provided Methods

fn init_loading_widgets(_root: Self::Root) -> Option<LoadingWidgets>

Allows you to initialize the root widget with a temporary value as a placeholder until the init_model() future completes.

This method does nothing by default.

fn update( &mut self, message: Self::Input, sender: AsyncFactorySender<Self>, ) -> impl Future<Output = ()>

Processes inputs received by the component.

fn update_cmd( &mut self, message: Self::CommandOutput, sender: AsyncFactorySender<Self>, ) -> impl Future<Output = ()>

Defines how the component should respond to command updates.

fn update_cmd_with_view( &mut self, widgets: &mut Self::Widgets, message: Self::CommandOutput, sender: AsyncFactorySender<Self>, ) -> impl Future<Output = ()>

Handles updates from a command.

fn update_view( &self, widgets: &mut Self::Widgets, sender: AsyncFactorySender<Self>, )

Updates the view after the model has been updated.

fn update_with_view( &mut self, widgets: &mut Self::Widgets, message: Self::Input, sender: AsyncFactorySender<Self>, ) -> impl Future<Output = ()>

Updates the model and view. Optionally returns a command to run.

fn shutdown( &mut self, widgets: &mut Self::Widgets, output: Sender<Self::Output>, )

Last method called before a factory component is shut down.

fn id(&self) -> String

An identifier for the component used for debug logging.

The default implementation of this method uses the address of the component, but implementations are free to provide more meaningful identifiers.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.