Trait relm4::factory::FactoryView

pub trait FactoryView<Widget> {
    type Root: Debug;
    type Position;

    // Required methods
    fn add(&self, widget: &Widget, position: &Self::Position) -> Self::Root;
    fn remove(&self, widget: &Self::Root);
}

A trait implemented for GTK4 widgets that allows a factory to create and remove widgets.

Required Associated Types

type Root: Debug

Widget type that’s stored inside a factory data type.

type Position

Position type used by this widget.

For example GridPosition for gtk::Grid or () for gtk::Box

Required Methods

fn add(&self, widget: &Widget, position: &Self::Position) -> Self::Root

Adds a new widget to self at the end.

fn remove(&self, widget: &Self::Root)

Removes a widget from self at the end.