Module relm4::factory::traits

Traits for for managing and updating factories.

Traits

FactoryComponent

A component that’s stored inside a factory. Similar to Component but adjusted to fit the life cycle of factories.

FactoryView

A trait implemented for GTK4 widgets that allows a factory to create and remove widgets.

Position

Returns the position of an element inside a container like gtk::Grid where the position isn’t clearly defined by the index.