Trait relm4::factory::traits::FactoryComponent

pub trait FactoryComponent<ParentWidget: FactoryView, ParentMsg>: Sized + Debug + 'static {
    type Command: 'static + Send;
    type CommandOutput: 'static + Send;
    type Input: 'static;
    type Output: 'static;
    type InitParams;
    type Root: Debug + OnDestroy + Clone;
    type Widgets: 'static;

    fn init_model(
        params: Self::InitParams, 
        index: &DynamicIndex, 
        input: &Sender<Self::Input>, 
        output: &Sender<Self::Output>
    ) -> Self;
    fn init_root(&self) -> Self::Root;
    fn init_widgets(
        &mut self, 
        index: &DynamicIndex, 
        root: &Self::Root, 
        returned_widget: &ParentWidget::ReturnedWidget, 
        input: &Sender<Self::Input>, 
        output: &Sender<Self::Output>
    ) -> Self::Widgets;

    fn output_to_parent_msg(_output: Self::Output) -> Option<ParentMsg> { ... }
    fn update(
        &mut self, 
        message: Self::Input, 
        input: &Sender<Self::Input>, 
        output: &Sender<Self::Output>
    ) -> Option<Self::Command> { ... }
    fn update_cmd(
        &mut self, 
        message: Self::CommandOutput, 
        input: &Sender<Self::Input>, 
        output: &Sender<Self::Output>
    ) { ... }
    fn update_cmd_with_view(
        &mut self, 
        widgets: &mut Self::Widgets, 
        message: Self::CommandOutput, 
        input: &Sender<Self::Input>, 
        output: &Sender<Self::Output>
    ) { ... }
    fn update_view(
        &self, 
        widgets: &mut Self::Widgets, 
        input: &Sender<Self::Input>, 
        output: &Sender<Self::Output>
    ) { ... }
    fn update_with_view(
        &mut self, 
        widgets: &mut Self::Widgets, 
        message: Self::Input, 
        input: &Sender<Self::Input>, 
        output: &Sender<Self::Output>
    ) -> Option<Self::Command> { ... }
    fn command(
        message: Self::Command, 
        shutdown: ShutdownReceiver, 
        output: Sender<Self::CommandOutput>
    ) -> CommandFuture { ... }
    fn shutdown(
        &mut self, 
        widgets: &mut Self::Widgets, 
        output: Sender<Self::Output>
    ) { ... }
}

A component that’s stored inside a factory. Similar to Component but adjusted to fit the life cycle of factories.

Required Associated Types

type Command: 'static + Send

Internal commands to perform

type CommandOutput: 'static + Send

Messages which are received from commands executing in the background.

type Input: 'static

The message type that the component accepts as inputs.

type Output: 'static

The message type that the component provides as outputs.

type InitParams

The initial parameter(s) for launch.

type Root: Debug + OnDestroy + Clone

The widget that was constructed by the component.

type Widgets: 'static

The type that’s used for storing widgets created for this component.

Required Methods

fn init_model(
params: Self::InitParams,
index: &DynamicIndex,
input: &Sender<Self::Input>,
output: &Sender<Self::Output>
) -> Self

Initializes the model.

fn init_root(&self) -> Self::Root

Initializes the root widget

fn init_widgets(
&mut self,
index: &DynamicIndex,
root: &Self::Root,
returned_widget: &ParentWidget::ReturnedWidget,
input: &Sender<Self::Input>,
output: &Sender<Self::Output>
) -> Self::Widgets

Initializes the widgets.

Provided Methods

fn output_to_parent_msg(_output: Self::Output) -> Option<ParentMsg>

Convert Self::Output into ParentMsg in order to send message to the parent.

If None is returned, nothing is forwarded.

fn update(
&mut self,
message: Self::Input,
input: &Sender<Self::Input>,
output: &Sender<Self::Output>
) -> Option<Self::Command>

Processes inputs received by the component.

fn update_cmd(
&mut self,
message: Self::CommandOutput,
input: &Sender<Self::Input>,
output: &Sender<Self::Output>
)

Defines how the component should respond to command updates.

fn update_cmd_with_view(
&mut self,
widgets: &mut Self::Widgets,
message: Self::CommandOutput,
input: &Sender<Self::Input>,
output: &Sender<Self::Output>
)

Handles updates from a command.

fn update_view(
&self,
widgets: &mut Self::Widgets,
input: &Sender<Self::Input>,
output: &Sender<Self::Output>
)

Updates the view after the model has been updated.

fn update_with_view(
&mut self,
widgets: &mut Self::Widgets,
message: Self::Input,
input: &Sender<Self::Input>,
output: &Sender<Self::Output>
) -> Option<Self::Command>

Updates the model and view. Optionally returns a command to run.

fn command(
message: Self::Command,
shutdown: ShutdownReceiver,
output: Sender<Self::CommandOutput>
) -> CommandFuture

A command to perform in a background thread.

fn shutdown(
&mut self,
widgets: &mut Self::Widgets,
output: Sender<Self::Output>
)

Last method called before a component is shut down.