Trait relm4::factory::traits::Position

pub trait Position<Pos> {
    fn position(index: usize) -> Pos;
}

Returns the position of an element inside a container like gtk::Grid where the position isn’t clearly defined by the index.

Required Methods

fn position(index: usize) -> Pos

Returns the position.

This function can be called very often if widgets are moved a lot, so it should be cheap to call.