relm4::factory::widgets

Module traits

Trait used for factories to interact with widgets.

Traits

AsyncPosition

Returns the position of an element inside a container like gtk::Grid where the position isn’t clearly defined by the index.

FactoryView

A trait implemented for GTK4 widgets that allows a factory to create and remove widgets.

Position

Returns the position of an element inside a container like gtk::Grid where the position isn’t clearly defined by the index.