relm4::factory::widgets::traits

Trait AsyncPosition

pub trait AsyncPosition<Pos> {
    // Required method
    fn position(index: usize) -> Pos;
}

Returns the position of an element inside a container like gtk::Grid where the position isn’t clearly defined by the index.

Unlike Position, this trait doesn’t get access to self, because the model might not be initialized when the widgets are updated in the factory.

Required Methods

fn position(index: usize) -> Pos

Returns the position.

This function can be called very often if widgets are moved a lot, so it should be cheap to call.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.