Function relm4::spawn_future

pub fn spawn_future<F: Future<Output = ()> + Send + 'static>(f: F)

Spawns a future on the main thread in the main event loop.

Panics

This function itself doesn’t panic but it might panic if you run futures that expect the tokio runtime. Use the tokio-rt feature and an AsyncComponent for this instead.

Examples found in repository

relm4-examples/examples/future.rs (line 109)

↕

89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118

fn update(&mut self, msg: AppMsg, _components: &(), sender: Sender<AppMsg>) -> bool {
        self.reset();

        match msg {
            AppMsg::Request(url) => {
                self.set_waiting(true);

                let fut = async move {
                    let mut text = "Connection error".to_string();

                    if surf::Url::parse(&url).is_ok() {
                        if let Ok(mut req) = surf::get(url).await {
                            if let Ok(resp) = req.body_string().await {
                                text = resp;
                            }
                        }
                    }
                    sender.send(AppMsg::Response(text)).unwrap();
                };

                spawn_future(fut);
            }
            AppMsg::Response(text) => {
                self.set_text(text);
                self.set_waiting(false);
            }
        }

        true
    }