relm4

Module prelude

Commonly-imported traits and types.

Modules that contain components can glob import this module to bring all needed types and traits into scope.

Re-exports

pub use crate::component::AsyncComponent;

pub use crate::component::AsyncComponentController;

pub use crate::component::AsyncComponentParts;

pub use crate::component::AsyncComponentSender;

pub use crate::component::AsyncController;

pub use crate::component::SimpleAsyncComponent;

pub use crate::factory::AsyncFactoryComponent;

pub use crate::factory::AsyncFactorySender;

pub use crate::factory::AsyncFactoryVecDeque;

pub use crate::factory::DynamicIndex;

pub use crate::factory::FactoryComponent;

pub use crate::factory::FactorySender;

pub use crate::factory::FactoryVecDeque;

pub use crate::Component;

pub use crate::ComponentController;

pub use crate::ComponentParts;

pub use crate::ComponentSender;

pub use crate::Controller;

pub use crate::RelmApp;

pub use crate::RelmWidgetExt;

pub use crate::SimpleComponent;

pub use crate::WidgetRef;

pub use crate::WidgetTemplate;

pub use adw;

pub use gtk;

pub use panel;