relm4

Module shared_state

Shared state that can be accessed by many components.

Structs

AsyncReducer

A type that allows you to share information across your application easily with async operations.

Reducer

A type that allows you to share information across your application easily.

SharedState

A type that allows you to share information across your application easily. Get immutable and mutable access to the data and subscribe to changes.

SharedStateReadGuard

A guard that immutably dereferences Data.

SharedStateWriteGuard

A guard that mutably dereferences Data. Once dropped all subscribers of the SharedState will be notified.

Traits

AsyncReducible

A trait that implements an async reducer function.

Reducible

A trait that implements a reducer function.