relm4::shared_state

Struct SharedState

pub struct SharedState<Data> { /* private fields */ }

A type that allows you to share information across your application easily. Get immutable and mutable access to the data and subscribe to changes.

Panics

SharedState uses a RwLock internally. If you use Self::read() and Self::write() in the same scope your code might be stuck in a deadlock or panic.

Implementations

impl<Data> SharedState<Data>
where Data: Default,

pub const fn new() -> Self

Create a new SharedState variable.

The data will be initialized lazily on the first access.

Example

use relm4::SharedState;
static STATE: SharedState<MyData> = SharedState::new();

Examples found in repository?

relm4/examples/tab_game.rs (line 19)

19static GAME_STATE: SharedState<GameState> = SharedState::new();

pub fn subscribe<Msg, F>(&self, sender: &Sender<Msg>, f: F)
where F: Fn(&Data) -> Msg + 'static + Send + Sync, Msg: Send + 'static,

Subscribe to a shared state type. Any subscriber will be notified with a message every time you modify the shared state using Self::get_mut().

use relm4::SharedState;
static STATE: SharedState<u8> = SharedState::new();

let (sender, receiver) = relm4::channel();

// Every time we modify the data, we will receive
// the updated value as a message.
STATE.subscribe(&sender, |data| *data);

{
    let mut data = STATE.write();
    *data += 1;
}

assert_eq!(receiver.recv_sync().unwrap(), 1);

Examples found in repository?

relm4/examples/tab_game.rs (line 141)

140    fn init_model(value: Self::Init, _index: &DynamicIndex, sender: FactorySender<Self>) -> Self {
141        GAME_STATE.subscribe(sender.input_sender(), |_| CounterMsg::Update);
142        Self { id: value }
143    }

pub fn subscribe_optional<Msg, F>(&self, sender: &Sender<Msg>, f: F)
where F: Fn(&Data) -> Option<Msg> + 'static + Send + Sync, Msg: Send + 'static,

An alternative version of subscribe() that only send a message if the closure returns Some.

pub fn read(&self) -> SharedStateReadGuard<'_, Data>

Get immutable access to the shared data.

Returns a RAII guard which will release this thread’s shared access once it is dropped.

The calling thread will be blocked until there are no more writers which hold the lock (see RwLock).

Panics

This function will panic if the internal RwLock is poisoned. A RwLock is poisoned whenever a writer panics while holding an exclusive lock. The failure will occur immediately after the lock has been acquired.

Also, this function might panic when called if the lock is already held by the current thread.

Examples found in repository?

relm4/examples/tab_game.rs (line 152)

145    fn init_widgets(
146        &mut self,
147        index: &DynamicIndex,
148        root: Self::Root,
149        returned_widget: &adw::TabPage,
150        sender: FactorySender<Self>,
151    ) -> Self::Widgets {
152        let state = GAME_STATE.read();
153        let widgets = view_output!();
154        widgets
155    }

pub fn try_read( &self, ) -> Result<SharedStateReadGuard<'_, Data>, TryLockError<RwLockReadGuard<'_, Data>>>

Get immutable access to the shared data.

Similar to read, but doesn’t block so this function simply returns an Err if the data is already locked (or poisoned).

pub fn write(&self) -> SharedStateWriteGuard<'_, Data>

Get mutable access to the shared data.

Returns a RAII guard which will notify all subscribers and release this thread’s shared access once it is dropped.

This function will not return while other writers or other readers currently have access to the internal lock (see RwLock).

Panics

This function will panic if the internal RwLock is poisoned. A RwLock is poisoned whenever a writer panics while holding an exclusive lock. The failure will occur immediately after the lock has been acquired.

Also, this function might panic when called if the lock is already held by the current thread.

Example

static STATE: SharedState<u8> = SharedState::new();

// Overwrite the current value with 1
*STATE.write() = 1;

Panic example

static STATE: SharedState<u8> = SharedState::new();

let read_guard = STATE.read();

// This is fine
let another_read_guard = STATE.read();

// This might panic or result in a dead lock
// because you cannot read and write at the same time.
// To solve this, drop all read guards on this thread first.
let another_write_guard = STATE.write();

Examples found in repository?

relm4/examples/tab_game.rs (line 252)

246    fn update(&mut self, msg: Self::Input, sender: ComponentSender<Self>, _root: &Self::Root) {
247        match msg {
248            AppMsg::StartGame(index) => {
249                self.start_index = Some(index);
250                sender.command(|sender, _| async move {
251                    for i in (1..4).rev() {
252                        *GAME_STATE.write() = GameState::Countdown(i);
253                        relm4::tokio::time::sleep(Duration::from_millis(1000)).await;
254                    }
255                    *GAME_STATE.write() = GameState::Running;
256                    for _ in 0..20 {
257                        relm4::tokio::time::sleep(Duration::from_millis(500)).await;
258                        sender.send(false).unwrap();
259                    }
260                    relm4::tokio::time::sleep(Duration::from_millis(1000)).await;
261                    sender.send(true).unwrap();
262                });
263            }
264            AppMsg::StopGame => {
265                *GAME_STATE.write() = GameState::Guessing;
266            }
267            AppMsg::SelectedGuess(index) => {
268                *GAME_STATE.write() = GameState::End(index == self.start_index.take().unwrap());
269            }
270        }
271    }

pub fn try_write( &self, ) -> Result<SharedStateWriteGuard<'_, Data>, TryLockError<RwLockWriteGuard<'_, Data>>>

Get mutable access to the shared data.

Similar to write, but doesn’t block so this function simply returns an Err if the data is already locked (or poisoned).

pub fn get_mut(&mut self) -> &mut Data

Get mutable access to the shared data. Since this call borrows the SharedState mutably, no actual locking needs to take place, but the mutable borrow statically guarantees no locks exist.

This method will not notify any subscribers!

Panics

This function will panic if the internal RwLock is poisoned. A RwLock is poisoned whenever a writer panics while holding an exclusive lock. The failure will occur immediately after the lock has been acquired.

pub fn read_inner(&self) -> RwLockReadGuard<'_, Data>

Get immutable access to the shared data.

This method will not notify any subscribers!

Panics

This function will panic if the internal RwLock is poisoned. A RwLock is poisoned whenever a writer panics while holding an exclusive lock. The failure will occur immediately after the lock has been acquired.

Also, this function might panic when called if the lock is already held by the current thread.

pub fn write_inner(&self) -> RwLockWriteGuard<'_, Data>

Get mutable access to the shared data.

This method will not notify any subscribers!

Panics

This function will panic if the internal RwLock is poisoned. A RwLock is poisoned whenever a writer panics while holding an exclusive lock. The failure will occur immediately after the lock has been acquired.

Also, this function might panic when called if the lock is already held by the current thread.