relm4::shutdown

Function channel

pub fn channel() -> (ShutdownSender, ShutdownReceiver)

Creates a broadcasting shutdown channel.

The sending side is responsible for initiating a shutdown. The receiving side is responsible for responding to shutdowns.