relm4

Struct RelmApp

pub struct RelmApp<M: Debug + 'static> { /* private fields */ }

An app that runs the main application.

Implementations

impl<M: Debug + 'static> RelmApp<M>

pub fn new(app_id: &str) -> Self

Create a new Relm4 application.

This function will create a new gtk::Application object if necessary.

If the libadwaita feature is enabled, then the created gtk::Application will be an instance of adw::Application. This can be overridden by passing your own application object to RelmApp::from_app.

Examples found in repository?

relm4/examples/menu.rs (line 168)

167fn main() {
168    let app = RelmApp::new("relm4.example.menu");
169    app.run::<App>(0);
170}

Hide additional examples

relm4-components/examples/alert.rs (line 153)

152fn main() {
153    let app = RelmApp::new("relm4.example.alert");
154    app.run::<App>(());
155}

relm4/examples/data_binding.rs (line 102)

101fn main() {
102    let app = RelmApp::new("relm4.example.simple");
103    app.run::<App>(0);
104}

relm4/examples/entry.rs (line 132)

131fn main() {
132    let app = RelmApp::new("relm4.example.entry");
133    app.run::<App>(());
134}

relm4/examples/to_do.rs (line 162)

161fn main() {
162    let app = RelmApp::new("relm4.example.to_do");
163    app.run::<App>(());
164}

relm4/examples/toast.rs (line 108)

107fn main() {
108    let app = RelmApp::new("relm4.example.toast");
109    app.run::<App>(());
110}

Additional examples can be found in:

relm4/examples/factory.rs

relm4/examples/factory_async.rs

relm4/examples/factory_hash_map.rs

relm4/examples/worker.rs

relm4/examples/actions.rs

relm4/examples/popover.rs

relm4/examples/drawing.rs

relm4/examples/tab_game.rs

relm4-components/examples/combo_box.rs

relm4/examples/components.rs

relm4/examples/simple.rs

relm4/examples/tab_factory.rs

relm4-components/examples/open_button.rs

relm4-components/examples/web_image.rs

relm4/examples/grid_factory.rs

relm4/examples/multi_window.rs

relm4-components/examples/file_dialogs.rs

relm4/examples/embedded_logo.rs

relm4-components/examples/adw_combo_row.rs

relm4/examples/leaflet_sidebar.rs

relm4/examples/message_broker.rs

relm4/examples/message_stream.rs

relm4/examples/typed_grid_view.rs

relm4/examples/typed_list_view.rs

relm4/examples/transient_dialog.rs

relm4/examples/typed_column_view.rs

relm4/examples/progress.rs

relm4/examples/simple_manual.rs

relm4/examples/split_layout.rs

relm4/examples/state_management.rs

relm4/examples/drag_and_drop.rs

relm4/examples/non_blocking_async.rs

relm4/examples/non_blocking_sync.rs

relm4/examples/drop_sub_components.rs

relm4/examples/widget_template.rs

relm4/examples/typed_list_view_async.rs

relm4/examples/simple_async.rs

relm4/examples/navigation_splitview_with_stack.rs

relm4/examples/macro_reference.rs

relm4/examples/tracker.rs

relm4/examples/log.rs

pub fn from_app(app: impl IsA<Application>) -> Self

Create a Relm4 application with a provided gtk::Application.

pub fn with_broker(self, broker: &'static MessageBroker<M>) -> Self

Add MessageBroker to the top-level component.

pub fn with_args(self, args: Vec<String>) -> Self

Add command line arguments to run with.

pub fn visible_on_activate(self, visible: bool) -> Self

If true, make the window visible whenever the app is activated (e. g. every time RelmApp::run is called).

By default, this value is true. If you don’t want the window to be visible immediately (especially when using async components), you can set this to false and call WidgetExt::set_visible() manually on your window.

pub fn allow_multiple_instances(&self, allow: bool)

If true, allow multiple concurrent instances of the application by setting the gtk::gio::ApplicationFlags::NON_UNIQUE flag.

By default, this flag is not set. When the flag is not set, the application will not be started if another instance is already running.

pub fn set_global_css_with_priority(&self, style_data: &str, priority: u32)

👎Deprecated: Use relm4::set_global_css_with_priority instead

Sets a custom global stylesheet, with the given priority.

The priority can be any value, but GTK includes some that you can use.

pub fn set_global_css(&self, style_data: &str)

👎Deprecated: Use relm4::set_global_css instead

Sets a custom global stylesheet.

pub fn set_global_css_from_file_with_priority<P: AsRef<Path>>( &self, path: P, priority: u32, ) -> Result<(), Error>

👎Deprecated: Use relm4::set_global_css_from_file_with_priority instead

Sets a custom global stylesheet from a file, with the given priority.

If the file doesn’t exist a tracing::error message will be emitted and an std::io::Error will be returned.

The priority can be any value, but GTK includes some that you can use.

pub fn set_global_css_from_file<P: AsRef<Path>>( &self, path: P, ) -> Result<(), Error>

👎Deprecated: Use relm4::set_global_css_from_file instead

Sets a custom global stylesheet from a file.

If the file doesn’t exist a tracing::error message will be emitted and an std::io::Error will be returned.

pub fn run<C>(self, payload: C::Init)
where C: Component<Input = M>, C::Root: AsRef<Window>,

Runs the application, returns once the application is closed.

Examples found in repository?

relm4/examples/menu.rs (line 169)

167fn main() {
168    let app = RelmApp::new("relm4.example.menu");
169    app.run::<App>(0);
170}

Hide additional examples

relm4-components/examples/alert.rs (line 154)

152fn main() {
153    let app = RelmApp::new("relm4.example.alert");
154    app.run::<App>(());
155}

relm4/examples/data_binding.rs (line 103)

101fn main() {
102    let app = RelmApp::new("relm4.example.simple");
103    app.run::<App>(0);
104}

relm4/examples/entry.rs (line 133)

131fn main() {
132    let app = RelmApp::new("relm4.example.entry");
133    app.run::<App>(());
134}

relm4/examples/to_do.rs (line 163)

161fn main() {
162    let app = RelmApp::new("relm4.example.to_do");
163    app.run::<App>(());
164}

relm4/examples/toast.rs (line 109)

107fn main() {
108    let app = RelmApp::new("relm4.example.toast");
109    app.run::<App>(());
110}

Additional examples can be found in:

relm4/examples/factory.rs

relm4/examples/factory_async.rs

relm4/examples/factory_hash_map.rs

relm4/examples/worker.rs

relm4/examples/actions.rs

relm4/examples/popover.rs

relm4/examples/drawing.rs

relm4/examples/tab_game.rs

relm4-components/examples/combo_box.rs

relm4/examples/components.rs

relm4/examples/simple.rs

relm4/examples/tab_factory.rs

relm4-components/examples/open_button.rs

relm4-components/examples/web_image.rs

relm4/examples/grid_factory.rs

relm4/examples/multi_window.rs

relm4-components/examples/file_dialogs.rs

relm4/examples/embedded_logo.rs

relm4-components/examples/adw_combo_row.rs

relm4/examples/leaflet_sidebar.rs

relm4/examples/message_broker.rs

relm4/examples/message_stream.rs

relm4/examples/typed_grid_view.rs

relm4/examples/typed_list_view.rs

relm4/examples/transient_dialog.rs

relm4/examples/typed_column_view.rs

relm4/examples/progress.rs

relm4/examples/simple_manual.rs

relm4/examples/split_layout.rs

relm4/examples/state_management.rs

relm4/examples/drag_and_drop.rs

relm4/examples/non_blocking_async.rs

relm4/examples/non_blocking_sync.rs

relm4/examples/drop_sub_components.rs

relm4/examples/widget_template.rs

relm4/examples/typed_list_view_async.rs

relm4/examples/navigation_splitview_with_stack.rs

relm4/examples/macro_reference.rs

relm4/examples/tracker.rs

relm4/examples/log.rs

pub fn run_async<C>(self, payload: C::Init)
where C: AsyncComponent<Input = M>, C::Root: AsRef<Window>,

Runs the application, returns once the application is closed.

Examples found in repository?

relm4/examples/simple_async.rs (line 109)

107fn main() {
108    let app = RelmApp::new("relm4.example.simple_async");
109    app.run_async::<AppModel>(0);
110}