Trait relm4::Component

pub trait Component: Sized + 'static {
    type CommandOutput: 'static + Send;
    type Input: 'static;
    type Output: 'static;
    type InitParams;
    type Root: Debug + OnDestroy;
    type Widgets: 'static;

    fn init_root() -> Self::Root;
    fn init(
        params: Self::InitParams, 
        root: &Self::Root, 
        sender: &ComponentSender<Self>
    ) -> ComponentParts<Self>;

    fn builder() -> ComponentBuilder<Self> { ... }
    fn update(&mut self, message: Self::Input, sender: &ComponentSender<Self>) { ... }
    fn update_cmd(
        &mut self, 
        message: Self::CommandOutput, 
        sender: &ComponentSender<Self>
    ) { ... }
    fn update_cmd_with_view(
        &mut self, 
        widgets: &mut Self::Widgets, 
        message: Self::CommandOutput, 
        sender: &ComponentSender<Self>
    ) { ... }
    fn update_view(
        &self, 
        widgets: &mut Self::Widgets, 
        sender: &ComponentSender<Self>
    ) { ... }
    fn update_with_view(
        &mut self, 
        widgets: &mut Self::Widgets, 
        message: Self::Input, 
        sender: &ComponentSender<Self>
    ) { ... }
    fn shutdown(
        &mut self, 
        widgets: &mut Self::Widgets, 
        output: Sender<Self::Output>
    ) { ... }
}

Elm-style variant of a Component with view updates separated from input updates

Required Associated Types

type CommandOutput: 'static + Send

Messages which are received from commands executing in the background.

type Input: 'static

The message type that the component accepts as inputs.

type Output: 'static

The message type that the component provides as outputs.

type InitParams

The initial parameter(s) for launch.

type Root: Debug + OnDestroy

The widget that was constructed by the component.

type Widgets: 'static

The type that’s used for storing widgets created for this component.

Required Methods

fn init_root() -> Self::Root

Initializes the root widget

fn init(
params: Self::InitParams,
root: &Self::Root,
sender: &ComponentSender<Self>
) -> ComponentParts<Self>

Creates the initial model and view, docking it into the component.

Provided Methods

fn builder() -> ComponentBuilder<Self>

Create a builder for this component.

Examples found in repository?

examples/settings_list.rs (line 24)

↕

19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69

fn main() {
    gtk::Application::builder()
        .application_id("org.relm4.SettingsListExample")
        .launch(|_app, window| {
            // Intiialize a component's root widget
            let component = App::builder()
                // Attach the root widget to the given window.
                .attach_to(&window)
                // Start the component service with an initial parameter
                .launch("Settings List Demo".into())
                // Attach the returned receiver's messages to this closure.
                .connect_receiver(move |sender, message| match message {
                    Output::Clicked(id) => {
                        eprintln!("ID {id} Clicked");

                        match id {
                            0 => xdg_open("https://github.com/AaronErhardt/Relm4".into()),
                            1 => {
                                xdg_open("https://aaronerhardt.github.io/docs/relm4/relm4/".into())
                            }
                            2 => {
                                sender.send(Input::Clear);
                            }
                            _ => (),
                        }
                    }

                    Output::Reload => {
                        sender.send(Input::AddSetting {
                            description: "Browse GitHub Repository".into(),
                            button: "GitHub".into(),
                            id: 0,
                        });

                        sender.send(Input::AddSetting {
                            description: "Browse Documentation".into(),
                            button: "Docs".into(),
                            id: 1,
                        });

                        sender.send(Input::AddSetting {
                            description: "Clear List".into(),
                            button: "Clear".into(),
                            id: 2,
                        });
                    }
                });

            println!("parent is {:?}", component.widget().toplevel_window());
        });
}

fn update(&mut self, message: Self::Input, sender: &ComponentSender<Self>)

Processes inputs received by the component.

fn update_cmd(
&mut self,
message: Self::CommandOutput,
sender: &ComponentSender<Self>
)

Defines how the component should respond to command updates.

fn update_cmd_with_view(
&mut self,
widgets: &mut Self::Widgets,
message: Self::CommandOutput,
sender: &ComponentSender<Self>
)

Handles updates from a command.

fn update_view(
&self,
widgets: &mut Self::Widgets,
sender: &ComponentSender<Self>
)

Updates the view after the model has been updated.

fn update_with_view(
&mut self,
widgets: &mut Self::Widgets,
message: Self::Input,
sender: &ComponentSender<Self>
)

Updates the model and view. Optionally returns a command to run.

fn shutdown(
&mut self,
widgets: &mut Self::Widgets,
output: Sender<Self::Output>
)

Last method called before a component is shut down.