Trait relm4::MessageHandler

pub trait MessageHandler<ParentModel: Model> {
    type Msg;
    type Sender;
    fn init(
        parent_model: &ParentModel, 
        parent_sender: Sender<ParentModel::Msg>
    ) -> Self;
    fn send(&self, msg: Self::Msg);
    fn sender(&self) -> Self::Sender;
}

A message handler that can be used in situations where a RelmWorker isn’t flexible enough.

Associated Types

type Msg

The message type of this message handler.

type Sender

The sender type that can be used to send a message to a RelmMsgHandler.

Required methods

fn init(
parent_model: &ParentModel,
parent_sender: Sender<ParentModel::Msg>
) -> Self

Initialize this message handler.

fn send(&self, msg: Self::Msg)

Sends a message to the message handler.

fn sender(&self) -> Self::Sender

Get a sender for sending messages to this RelmMsgHandler.