Trait relm4::MicroWidgets

pub trait MicroWidgets<Model: MicroModel + ?Sized> {
    type Root: Debug;
    fn init_view(model: &Model, sender: Sender<Model::Msg>) -> Self;

    fn view(&mut self, model: &Model, sender: Sender<Model::Msg>);

    fn root_widget(&self) -> Self::Root;
}

Define behavior to turn the data of your MicroModel into widgets.

Associated Types

type Root: Debug

The root represents the first widget that all other widgets of this MicroComponent are attached to.

Required methods

fn init_view(model: &Model, sender: Sender<Model::Msg>) -> Self

Initialize the UI.

Use the sender to connect UI events and send messages back to modify the model.

fn view(&mut self, model: &Model, sender: Sender<Model::Msg>)

Update the view to represent the updated model.

fn root_widget(&self) -> Self::Root

Return a clone of the root widget. This is typically a GTK4 widget.