relm4

Trait RelmListBoxExt

pub trait RelmListBoxExt {
    // Required methods
    fn index_of_child(&self, widget: &impl AsRef<Widget>) -> Option<i32>;
    fn remove_row_of_child(&self, widget: &impl AsRef<Widget>);
    fn row_of_child(&self, widget: &impl AsRef<Widget>) -> Option<ListBoxRow>;
}

Additional methods for gtk::ListBox.

Required Methods

fn index_of_child(&self, widget: &impl AsRef<Widget>) -> Option<i32>

Get the index of a widget attached to a listbox.

fn remove_row_of_child(&self, widget: &impl AsRef<Widget>)

Remove the row of a child attached a listbox.

fn row_of_child(&self, widget: &impl AsRef<Widget>) -> Option<ListBoxRow>

Get the row of a widget attached to a listbox.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.