relm4

Trait RelmRemoveExt

pub trait RelmRemoveExt: ContainerChild {
    // Required method
    fn container_remove(&self, child: &impl AsRef<Self::Child>);
}

Widget types which can have widgets removed from them.

Required Methods

fn container_remove(&self, child: &impl AsRef<Self::Child>)

Removes the widget from the container if it is a child of the container.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.