relm4

Trait RelmSetChildExt

pub trait RelmSetChildExt: ContainerChild {
    // Required methods
    fn container_set_child(&self, widget: Option<&impl AsRef<Self::Child>>);
    fn container_get_child(&self) -> Option<Self::Child>;
}

Widget types which allow to set or unset their child.

Required Methods

fn container_set_child(&self, widget: Option<&impl AsRef<Self::Child>>)

Set a child for the container or remove it using None.

fn container_get_child(&self) -> Option<Self::Child>

Get the child of a container. Returns None if the container has no child.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.