Trait relm4::SimpleComponent

pub trait SimpleComponent: Sized + 'static {
    type Input: 'static;
    type Output: 'static;
    type InitParams;
    type Root: Debug + OnDestroy;
    type Widgets: 'static;

    fn init_root() -> Self::Root;
    fn init(
        params: Self::InitParams, 
        root: &Self::Root, 
        sender: &ComponentSender<Self>
    ) -> ComponentParts<Self>;

    fn update(&mut self, message: Self::Input, sender: &ComponentSender<Self>) { ... }
    fn update_cmd(
        &mut self, 
        input: &Sender<Self::Input>, 
        output: &Sender<Self::Output>
    ) { ... }
    fn update_view(
        &self, 
        widgets: &mut Self::Widgets, 
        sender: &ComponentSender<Self>
    ) { ... }
    fn shutdown(
        &mut self, 
        widgets: &mut Self::Widgets, 
        output: Sender<Self::Output>
    ) { ... }
}

Elm-style variant of a Component with view updates separated from input updates

Required Associated Types

type Input: 'static

The message type that the component accepts as inputs.

type Output: 'static

The message type that the component provides as outputs.

type InitParams

The initial parameter(s) for launch.

type Root: Debug + OnDestroy

The widget that was constructed by the component.

type Widgets: 'static

The type that’s used for storing widgets created for this component.

Required Methods

fn init_root() -> Self::Root

Initializes the root widget

fn init(
params: Self::InitParams,
root: &Self::Root,
sender: &ComponentSender<Self>
) -> ComponentParts<Self>

Creates the initial model and view, docking it into the component.

Provided Methods

fn update(&mut self, message: Self::Input, sender: &ComponentSender<Self>)

Processes inputs received by the component.

fn update_cmd(
&mut self,
input: &Sender<Self::Input>,
output: &Sender<Self::Output>
)

Defines how the component should respond to command updates.

fn update_view(
&self,
widgets: &mut Self::Widgets,
sender: &ComponentSender<Self>
)

Updates the view after the model has been updated.

fn shutdown(
&mut self,
widgets: &mut Self::Widgets,
output: Sender<Self::Output>
)

Last method called before a component is shut down.