relm4::typed_view::list

Struct TypedListView

pub struct TypedListView<T, S> {
    pub view: ListView,
    pub selection_model: S,
    /* private fields */
}

A high-level wrapper around gio::ListStore, gtk::SignalListItemFactory and gtk::ListView.

TypedListView aims at keeping nearly the same functionality and flexibility of the raw bindings while introducing a more idiomatic and type-safe interface.

Fields
view: ListView

The internal list view.

selection_model: S

The internal selection model.

Implementations

impl<T, S> TypedListView<T, S>
where T: RelmListItem + Ord, S: RelmSelectionExt,

pub fn with_sorting() -> Self

Create a new TypedListView that sorts the items based on the Ord trait.

Examples found in repository?

relm4/examples/typed_list_view.rs (line 138)

131    fn init(
132        counter: Self::Init,
133        root: Self::Root,
134        sender: ComponentSender<Self>,
135    ) -> ComponentParts<Self> {
136        // Initialize the ListView wrapper
137        let mut list_view_wrapper: TypedListView<MyListItem, gtk::SingleSelection> =
138            TypedListView::with_sorting();
139
140        // Add a filter and disable it
141        list_view_wrapper.add_filter(|item| item.value % 2 == 0);
142        list_view_wrapper.set_filter_status(0, false);
143
144        let model = App {
145            counter,
146            list_view_wrapper,
147        };
148
149        let my_view = &model.list_view_wrapper.view;
150
151        let widgets = view_output!();
152
153        ComponentParts { model, widgets }
154    }

Hide additional examples

relm4/examples/typed_list_view_async.rs (line 179)

172    fn init(
173        counter: Self::Init,
174        root: Self::Root,
175        sender: ComponentSender<Self>,
176    ) -> ComponentParts<Self> {
177        // Initialize the ListView wrapper
178        let mut list_view_wrapper: TypedListView<MyListItem, gtk::SingleSelection> =
179            TypedListView::with_sorting();
180
181        // Add a filter and disable it
182        list_view_wrapper.add_filter(|item| item.value % 2 == 0);
183        list_view_wrapper.set_filter_status(0, false);
184
185        let model = App {
186            counter,
187            list_view_wrapper,
188        };
189
190        let my_view = &model.list_view_wrapper.view;
191
192        let widgets = view_output!();
193
194        ComponentParts { model, widgets }
195    }

impl<T, S> TypedListView<T, S>
where T: RelmListItem, S: RelmSelectionExt,

pub fn new() -> Self

Create a new, empty TypedListView.

pub fn add_filter<F: Fn(&T) -> bool + 'static>(&mut self, f: F)

Add a function to filter the stored items. Returning false will simply hide the item.

Note that several filters can be added on top of each other.

Examples found in repository?

relm4/examples/typed_list_view.rs (line 141)

131    fn init(
132        counter: Self::Init,
133        root: Self::Root,
134        sender: ComponentSender<Self>,
135    ) -> ComponentParts<Self> {
136        // Initialize the ListView wrapper
137        let mut list_view_wrapper: TypedListView<MyListItem, gtk::SingleSelection> =
138            TypedListView::with_sorting();
139
140        // Add a filter and disable it
141        list_view_wrapper.add_filter(|item| item.value % 2 == 0);
142        list_view_wrapper.set_filter_status(0, false);
143
144        let model = App {
145            counter,
146            list_view_wrapper,
147        };
148
149        let my_view = &model.list_view_wrapper.view;
150
151        let widgets = view_output!();
152
153        ComponentParts { model, widgets }
154    }

Hide additional examples

relm4/examples/typed_list_view_async.rs (line 182)

172    fn init(
173        counter: Self::Init,
174        root: Self::Root,
175        sender: ComponentSender<Self>,
176    ) -> ComponentParts<Self> {
177        // Initialize the ListView wrapper
178        let mut list_view_wrapper: TypedListView<MyListItem, gtk::SingleSelection> =
179            TypedListView::with_sorting();
180
181        // Add a filter and disable it
182        list_view_wrapper.add_filter(|item| item.value % 2 == 0);
183        list_view_wrapper.set_filter_status(0, false);
184
185        let model = App {
186            counter,
187            list_view_wrapper,
188        };
189
190        let my_view = &model.list_view_wrapper.view;
191
192        let widgets = view_output!();
193
194        ComponentParts { model, widgets }
195    }

pub fn filters_len(&self) -> usize

Returns the amount of filters that were added.

pub fn set_filter_status(&mut self, idx: usize, active: bool) -> bool

Set a certain filter as active or inactive.

Examples found in repository?

relm4/examples/typed_list_view.rs (line 142)

131    fn init(
132        counter: Self::Init,
133        root: Self::Root,
134        sender: ComponentSender<Self>,
135    ) -> ComponentParts<Self> {
136        // Initialize the ListView wrapper
137        let mut list_view_wrapper: TypedListView<MyListItem, gtk::SingleSelection> =
138            TypedListView::with_sorting();
139
140        // Add a filter and disable it
141        list_view_wrapper.add_filter(|item| item.value % 2 == 0);
142        list_view_wrapper.set_filter_status(0, false);
143
144        let model = App {
145            counter,
146            list_view_wrapper,
147        };
148
149        let my_view = &model.list_view_wrapper.view;
150
151        let widgets = view_output!();
152
153        ComponentParts { model, widgets }
154    }
155
156    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
157        match msg {
158            Msg::Append => {
159                // Add 10 items
160                for _ in 0..10 {
161                    self.counter = self.counter.wrapping_add(1);
162                    self.list_view_wrapper.append(MyListItem::new(self.counter));
163                }
164
165                self.list_view_wrapper
166                    .iter()
167                    .for_each(|row| println!("item {}", row.borrow().value));
168
169                // Count up the first item
170                let first_item = self.list_view_wrapper.get(0).unwrap();
171                let first_binding = &mut first_item.borrow_mut().binding;
172                let mut guard = first_binding.guard();
173                *guard += 1;
174            }
175            Msg::Remove => {
176                // Remove the second item
177                self.list_view_wrapper.remove(1);
178            }
179            Msg::OnlyShowEven(show_only_even) => {
180                // Disable or enable the first filter
181                self.list_view_wrapper.set_filter_status(0, show_only_even);
182            }
183        }
184    }

Hide additional examples

relm4/examples/typed_list_view_async.rs (line 183)

172    fn init(
173        counter: Self::Init,
174        root: Self::Root,
175        sender: ComponentSender<Self>,
176    ) -> ComponentParts<Self> {
177        // Initialize the ListView wrapper
178        let mut list_view_wrapper: TypedListView<MyListItem, gtk::SingleSelection> =
179            TypedListView::with_sorting();
180
181        // Add a filter and disable it
182        list_view_wrapper.add_filter(|item| item.value % 2 == 0);
183        list_view_wrapper.set_filter_status(0, false);
184
185        let model = App {
186            counter,
187            list_view_wrapper,
188        };
189
190        let my_view = &model.list_view_wrapper.view;
191
192        let widgets = view_output!();
193
194        ComponentParts { model, widgets }
195    }
196
197    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
198        match msg {
199            Msg::Append => {
200                // Add 10 items
201                for _ in 0..10 {
202                    self.counter = self.counter.wrapping_add(1);
203                    self.list_view_wrapper.append(MyListItem::new(self.counter));
204                }
205            }
206            Msg::Remove => {
207                // Remove the second item
208                self.list_view_wrapper.remove(1);
209            }
210            Msg::OnlyShowEven(show_only_even) => {
211                // Disable or enable the first filter
212                self.list_view_wrapper.set_filter_status(0, show_only_even);
213            }
214        }
215    }

pub fn notify_filter_changed(&self, idx: usize) -> bool

Notify that a certain filter has changed. This causes the filter expression to be re-evaluated.

Returns true if a filter was notified.

pub fn pop_filter(&mut self)

Remove the last filter.

pub fn clear_filters(&mut self)

Remove all filters.

pub fn append(&mut self, value: T)

Add a new item at the end of the list.

Examples found in repository?

relm4/examples/typed_list_view_async.rs (line 203)

197    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
198        match msg {
199            Msg::Append => {
200                // Add 10 items
201                for _ in 0..10 {
202                    self.counter = self.counter.wrapping_add(1);
203                    self.list_view_wrapper.append(MyListItem::new(self.counter));
204                }
205            }
206            Msg::Remove => {
207                // Remove the second item
208                self.list_view_wrapper.remove(1);
209            }
210            Msg::OnlyShowEven(show_only_even) => {
211                // Disable or enable the first filter
212                self.list_view_wrapper.set_filter_status(0, show_only_even);
213            }
214        }
215    }

Hide additional examples

relm4/examples/typed_list_view.rs (line 162)

156    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
157        match msg {
158            Msg::Append => {
159                // Add 10 items
160                for _ in 0..10 {
161                    self.counter = self.counter.wrapping_add(1);
162                    self.list_view_wrapper.append(MyListItem::new(self.counter));
163                }
164
165                self.list_view_wrapper
166                    .iter()
167                    .for_each(|row| println!("item {}", row.borrow().value));
168
169                // Count up the first item
170                let first_item = self.list_view_wrapper.get(0).unwrap();
171                let first_binding = &mut first_item.borrow_mut().binding;
172                let mut guard = first_binding.guard();
173                *guard += 1;
174            }
175            Msg::Remove => {
176                // Remove the second item
177                self.list_view_wrapper.remove(1);
178            }
179            Msg::OnlyShowEven(show_only_even) => {
180                // Disable or enable the first filter
181                self.list_view_wrapper.set_filter_status(0, show_only_even);
182            }
183        }
184    }

pub fn extend_from_iter<I: IntoIterator<Item = T>>(&mut self, init: I)

Add new items from an iterator the the end of the list.

pub fn find<F: FnMut(&T) -> bool>(&self, equal_func: F) -> Option<u32>

Find the index of the first item that matches a certain function.

pub fn is_empty(&self) -> bool

Returns true if the list is empty.

pub fn len(&self) -> u32

Returns the length of the list (without filters).

pub fn get(&self, position: u32) -> Option<TypedListItem<T>>

Get the TypedListItem at the specified position.

Returns None if the position is invalid.

Examples found in repository?

relm4/examples/typed_list_view.rs (line 170)

156    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
157        match msg {
158            Msg::Append => {
159                // Add 10 items
160                for _ in 0..10 {
161                    self.counter = self.counter.wrapping_add(1);
162                    self.list_view_wrapper.append(MyListItem::new(self.counter));
163                }
164
165                self.list_view_wrapper
166                    .iter()
167                    .for_each(|row| println!("item {}", row.borrow().value));
168
169                // Count up the first item
170                let first_item = self.list_view_wrapper.get(0).unwrap();
171                let first_binding = &mut first_item.borrow_mut().binding;
172                let mut guard = first_binding.guard();
173                *guard += 1;
174            }
175            Msg::Remove => {
176                // Remove the second item
177                self.list_view_wrapper.remove(1);
178            }
179            Msg::OnlyShowEven(show_only_even) => {
180                // Disable or enable the first filter
181                self.list_view_wrapper.set_filter_status(0, show_only_even);
182            }
183        }
184    }

pub fn get_visible(&self, position: u32) -> Option<TypedListItem<T>>

Get the visible TypedListItem at the specified position, (the item at the given position after filtering and sorting).

Returns None if the position is invalid.

pub fn insert(&mut self, position: u32, value: T)

Insert an item at a specific position.

pub fn insert_sorted<F: FnMut(&T, &T) -> Ordering>( &self, value: T, compare_func: F, ) -> u32

Insert an item into the list and calculate its position from a sorting function.

pub fn remove(&mut self, position: u32)

Remove an item at a specific position.

Examples found in repository?

relm4/examples/typed_list_view_async.rs (line 208)

197    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
198        match msg {
199            Msg::Append => {
200                // Add 10 items
201                for _ in 0..10 {
202                    self.counter = self.counter.wrapping_add(1);
203                    self.list_view_wrapper.append(MyListItem::new(self.counter));
204                }
205            }
206            Msg::Remove => {
207                // Remove the second item
208                self.list_view_wrapper.remove(1);
209            }
210            Msg::OnlyShowEven(show_only_even) => {
211                // Disable or enable the first filter
212                self.list_view_wrapper.set_filter_status(0, show_only_even);
213            }
214        }
215    }

Hide additional examples

relm4/examples/typed_list_view.rs (line 177)

156    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
157        match msg {
158            Msg::Append => {
159                // Add 10 items
160                for _ in 0..10 {
161                    self.counter = self.counter.wrapping_add(1);
162                    self.list_view_wrapper.append(MyListItem::new(self.counter));
163                }
164
165                self.list_view_wrapper
166                    .iter()
167                    .for_each(|row| println!("item {}", row.borrow().value));
168
169                // Count up the first item
170                let first_item = self.list_view_wrapper.get(0).unwrap();
171                let first_binding = &mut first_item.borrow_mut().binding;
172                let mut guard = first_binding.guard();
173                *guard += 1;
174            }
175            Msg::Remove => {
176                // Remove the second item
177                self.list_view_wrapper.remove(1);
178            }
179            Msg::OnlyShowEven(show_only_even) => {
180                // Disable or enable the first filter
181                self.list_view_wrapper.set_filter_status(0, show_only_even);
182            }
183        }
184    }

pub fn clear(&mut self)

Remove all items.

pub fn iter(&self) -> TypedIterator<'_, TypedListView<T, S>>

Returns an iterator that allows modifying each TypedListItem.

Examples found in repository?

relm4/examples/typed_list_view.rs (line 166)

156    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
157        match msg {
158            Msg::Append => {
159                // Add 10 items
160                for _ in 0..10 {
161                    self.counter = self.counter.wrapping_add(1);
162                    self.list_view_wrapper.append(MyListItem::new(self.counter));
163                }
164
165                self.list_view_wrapper
166                    .iter()
167                    .for_each(|row| println!("item {}", row.borrow().value));
168
169                // Count up the first item
170                let first_item = self.list_view_wrapper.get(0).unwrap();
171                let first_binding = &mut first_item.borrow_mut().binding;
172                let mut guard = first_binding.guard();
173                *guard += 1;
174            }
175            Msg::Remove => {
176                // Remove the second item
177                self.list_view_wrapper.remove(1);
178            }
179            Msg::OnlyShowEven(show_only_even) => {
180                // Disable or enable the first filter
181                self.list_view_wrapper.set_filter_status(0, show_only_even);
182            }
183        }
184    }