Trait relm4::util::default_widgets::DefaultWidget

pub trait DefaultWidget {
    fn default_widget() -> Self;
}

This is mainly useful for macros that generate widgets in order to avoid passing arguments to new(...) functions of GTK widgets.

Required methods

fn default_widget() -> Self

Returns a widget with default configuration