Trait relm4::util::WidgetPlus

pub trait WidgetPlus {
    fn set_margin_all(&self, margin: i32);
    fn set_class_active(&self, class: &str, active: bool);
    fn inline_css(&self, style: &str);
    fn try_remove(&self, widget: &impl IsA<Widget>) -> bool;
}

Trait that extends gtk::prelude::WidgetExt.

This trait’s main goal is to reduce redundant code and to provide helpful methods for the widgets macro of relm4-macros.

Required Methods

fn set_margin_all(&self, margin: i32)

Set margin at start, end, top and bottom all at once.

fn set_class_active(&self, class: &str, active: bool)

Add class name if active is true and remove class name if active is false

fn inline_css(&self, style: &str)

Add inline CSS instructions to a widget.

widget.inline_css("border: 1px solid red");

fn try_remove(&self, widget: &impl IsA<Widget>) -> bool

Try to remove a widget from a widget.

Returns true if the removal was successful and false if nothing was done.