Trait relm4::worker::Worker

pub trait Worker: Sized + Send + 'static {
    type InitParams: 'static + Send;
    type Input: 'static + Send + Debug;
    type Output: 'static + Send + Debug;

    fn init(params: Self::InitParams, sender: &ComponentSender<Self>) -> Self;
    fn update(&mut self, message: Self::Input, sender: &ComponentSender<Self>);
}

Receives inputs and outputs in the background.

Required Associated Types

type InitParams: 'static + Send

The initial parameters that will be used to build the worker state.

type Input: 'static + Send + Debug

The type of inputs that this worker shall receive.

type Output: 'static + Send + Debug

The typue of outputs that this worker shall send.

Required Methods

fn init(params: Self::InitParams, sender: &ComponentSender<Self>) -> Self

Defines the initial state of the worker.

fn update(&mut self, message: Self::Input, sender: &ComponentSender<Self>)

Defines how inputs will bep processed