Crate relm4

An idiomatic GUI library inspired by Elm and based on gtk4-rs.

This document was built at Friday June 10, 02:53 UTC.

Re-exports

pub use self::component::ComponentParts;

pub use self::component::OnDestroy;

pub use self::shared_state::SharedState;

pub use self::worker::Worker;

pub use self::worker::WorkerController;

pub use self::worker::WorkerHandle;

pub use gtk;

pub use adw;

pub use panel;

pub use tokio;

Modules

actions

Action utility.

component

Components are smaller mostly independent parts of your application.

drawing

Utility to help drawing on a gtk::DrawingArea in a Relm4 application. Create a DrawHandler, initialize it, and get its context when handling a message (that could be sent from the draw signal).

factory

Defines traits and data types to generate widgets from collections efficiently.

shared_state

Shared state that can be accessed by many components.

shutdown

Cancellation mechanism used by Relm4.

util

Utility traits for working with GTK widgets.

worker

A simpler version of components that does work in the background.

Macros

menu

A macro to create menus.

new_action_group

Create a new type that implements ActionGroupName.

new_stateful_action

Create a new type that implements ActionName with state and target type.

new_stateless_action

Create a new type that implements ActionName without state or target type.

sendDeprecated

A short macro for conveniently sending messages.

view

The view! macro allows you to construct your UI easily and cleanly.

Structs

ComponentBuilder

A component that is ready for docking and launch.

JoinHandle

An owned permission to join on a task (await its termination).

Receiver

A Relm4 receiver receives messages from a component or worker.

RelmApp

An app that runs the main application.

Sender

A Relm4 sender sends messages to a component or worker.

ShutdownReceiver

Listens to shutdown signals and constructs shutdown futures.

Statics

RELM_BLOCKING_THREADS

Defines the maximum number of background threads to spawn for handling blocking tasks.

RELM_THREADS

Defines how many threads that Relm should use for background tasks.

Traits

ApplicationBuilderExt

Additional methods for gtk::builders::ApplicationBuilder

Component

Elm-style variant of a Component with view updates separated from input updates

ComponentController

Shared behavior of component controller types.

ContainerChild

Type of children inside a container.

RelmContainerExt

Widget types which can have widgets attached to them.

RelmIterChildrenExt

Widget types which allow iteration over their children.

RelmListBoxExt

Additional methods for gtk::ListBox.

RelmRemoveAllExt

Widget types that allow removal of all their children.

RelmRemoveExt

Widget types which can have widgets removed from them.

RelmSetChildExt

Widget types which allow to set or unset their child.

RelmWidgetExt

Additional methods for gtk::Widget

SimpleComponent

Elm-style variant of a Component with view updates separated from input updates

WidgetPlus

Trait that extends gtk::prelude::WidgetExt.

WidgetRef

Get a reference to a widget.

Functions

channel

Create an unbounded channel to send messages between different parts of you application.

set_global_css

Sets a custom global stylesheet.

set_global_css_from_file

Sets a custom global stylesheet from a file.

spawn

Spawns a Send-able future to the shared component runtime.

spawn_blocking

Spawns a blocking task in a background thread pool

spawn_local

Spawns a thread-local future on GLib’s executor, for non-Send futures.

Type Definitions

ComponentSender

Contain senders used by the component.

Attribute Macros

component

Macro that implements relm4::Component and generates the corresponding struct.

factory