Crate relm4

An idiomatic GUI library inspired by Elm and based on gtk4-rs.

The docs are available in two versions. Use the stable docs if you want get information about a version that was already published. Visit the nightly docs if are trying out the newest but possibly unstable version of the crate.

Docs of related crates: relm4 | relm4-macros | relm4-components | gtk4-rs | gtk-rs-core | libadwaita-rs | libpanel-rs

GitHub | Website | Book | Blog

Re-exports

pub use component::worker::Worker;

pub use component::worker::WorkerController;

pub use component::worker::WorkerHandle;

pub use component::Component;

pub use component::ComponentBuilder;

pub use component::ComponentController;

pub use component::ComponentParts;

pub use component::Controller;

pub use component::MessageBroker;

pub use component::SimpleComponent;

pub use shared_state::Reducer;

pub use shared_state::Reducible;

pub use shared_state::SharedState;

pub use gtk;

pub use adw;

pub use panel;

pub use async_trait;

pub use once_cell;

pub use tokio;

Modules

actions

Action utility.

component

Components are smaller mostly independent parts of your application.

drawing

Utility to help drawing on a gtk::DrawingArea in a Relm4 application. Create a DrawHandler, initialize it, and get its context when handling a message (that could be sent from the draw signal).

factory

Defines traits and data types to generate widgets from collections efficiently.

loading_widgets

Utilities for removing temporary widgets from async factories or components.

prelude

Commonly-imported traits and types.

shared_state

Shared state that can be accessed by many components.

shutdown

Cancellation mechanism used by Relm4.

Macros

menu

A macro to create menus.

new_action_group

Create a new type that implements ActionGroupName.

new_stateful_action

Create a new type that implements ActionName with state and target type.

new_stateless_action

Create a new type that implements ActionName without state or target type.

sendDeprecated

A short macro for conveniently sending messages.

view

The view! macro allows you to construct your UI easily and cleanly.

Structs

AsyncComponentSender

Contains senders to send and receive messages from a Component.

AsyncFactorySender

Contains senders to send and receive messages from a Component.

ComponentSender

Contains senders to send and receive messages from a Component.

FactorySender

Contains senders to send and receive messages from a Component.

JoinHandle

An owned permission to join on a task (await its termination).

Receiver

A Relm4 receiver receives messages from a component or worker.

RelmApp

An app that runs the main application.

Sender

A Relm4 sender sends messages to a component or worker.

ShutdownReceiver

Listens to shutdown signals and constructs shutdown futures.

Statics

RELM_BLOCKING_THREADS

Defines the maximum number of background threads to spawn for handling blocking tasks.

RELM_THREADS

Defines how many threads that Relm4 should use for background tasks.

Traits

ApplicationBuilderExt

Additional methods for gtk::builders::ApplicationBuilder

ContainerChild

Type of children inside a container.

RelmContainerExt

Widget types which can have widgets attached to them.

RelmIterChildrenExt

Widget types which allow iteration over their children.

RelmListBoxExt

Additional methods for gtk::ListBox.

RelmObjectExt

Trait that extends gtk::prelude::ObjectExt.

RelmRemoveAllExt

Widget types that allow removal of all their children.

RelmRemoveExt

Widget types which can have widgets removed from them.

RelmSetChildExt

Widget types which allow to set or unset their child.

RelmWidgetExt

Trait that extends gtk::prelude::WidgetExt.

WidgetRef

Get a reference to a widget.

WidgetTemplate

A trait that describes a widget template.

Functions

channel

Create an unbounded channel to send messages between different parts of you application.

main_application

Returns the global gtk::Application that’s used internally by RelmApp.

set_global_css

Sets a custom global stylesheet.

set_global_css_from_file

Sets a custom global stylesheet from a file.

spawn

Spawns a Send-able future to the shared component runtime.

spawn_blocking

Spawns a blocking task in a background thread pool.

spawn_local

Spawns a thread-local future on GLib’s executor, for non-Send futures.

spawn_local_with_priority

Spawns a thread-local future on GLib’s executor, for non-Send futures.

Attribute Macros

component

Macro that implements relm4::Component or relm4::SimpleComponent and generates the corresponding widgets struct.

factory

Macro that implements relm4::factory::FactoryComponent and generates the corresponding widgets struct.

widget_template

A macro to generate widget templates.