Struct relm4::actions::RelmAction

pub struct RelmAction<Name: ActionName> { /* private fields */ }

A type safe action that wraps around gio::SimpleAction.

Implementations

impl<Name: ActionName> RelmAction<Name> where
Name::State: ToVariant + FromVariant,
Name::Target: ToVariant + FromVariant,

pub fn new_stateful_with_target_value<Callback: Fn(&SimpleAction, &mut Name::State, Name::Target) + 'static>(
start_value: &Name::State,
callback: Callback
) -> Self

Create a new stateful action with target value.

Examples found in repository?

examples/menu.rs (lines 126-129)

↕

82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140

    fn init(
        counter: Self::InitParams,
        root: &Self::Root,
        sender: &ComponentSender<Self>,
    ) -> ComponentParts<Self> {
        // ============================================================
        //
        // You can also use menu! outside of the widget macro.
        // This is the manual equivalent to the the menu! macro above.
        //
        // ============================================================
        //
        // relm4::menu! {
        //     main_menu: {
        //         custom: "my_widget",
        //         "Test" => TestAction,
        //         "Test2" => TestAction,
        //         "Test toggle" => TestU8Action(1_u8),
        //         section! {
        //             "Section test" => TestAction,
        //             "Test toggle" => TestU8Action(1_u8),
        //         },
        //         section! {
        //             "Test" => TestAction,
        //             "Test2" => TestAction,
        //             "Test Value" => TestU8Action(1_u8),
        //         }
        //     }
        // };

        let model = Self { counter };
        let widgets = view_output!();

        let group = RelmActionGroup::<WindowActionGroup>::new();

        let action: RelmAction<TestAction> = {
            let sender = sender.clone();
            RelmAction::new_stateless(move |_| {
                println!("Statelesss action!");
                sender.input(AppMsg::Increment);
            })
        };

        let action2: RelmAction<TestU8Action> =
            RelmAction::new_stateful_with_target_value(&0, |_, state, _value| {
                *state ^= 1;
                dbg!(state);
            });

        group.add_action(action);
        group.add_action(action2);

        let actions = group.into_action_group();
        widgets
            .main_window
            .insert_action_group("win", Some(&actions));

        ComponentParts { model, widgets }
    }

impl<Name: ActionName> RelmAction<Name> where
Name::State: ToVariant + FromVariant,
Name::Target: EmptyType,

pub fn new_stateful<Callback: Fn(&SimpleAction, &mut Name::State) + 'static>(
start_value: &Name::State,
callback: Callback
) -> Self

Create a new stateful action.

impl<Name: ActionName> RelmAction<Name> where
Name::State: EmptyType,
Name::Target: ToVariant + FromVariant,

pub fn new_with_target_value<Callback: Fn(&SimpleAction, Name::Target) + 'static>(
callback: Callback
) -> Self

Create a new stateless action with a target value.

impl<Name: ActionName> RelmAction<Name> where
Name::Target: EmptyType,
Name::State: EmptyType,

pub fn new_stateless<Callback: Fn(&SimpleAction) + 'static>(
callback: Callback
) -> Self

Create a new stateless action.

Examples found in repository?

examples/menu.rs (lines 119-122)

↕

82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140

    fn init(
        counter: Self::InitParams,
        root: &Self::Root,
        sender: &ComponentSender<Self>,
    ) -> ComponentParts<Self> {
        // ============================================================
        //
        // You can also use menu! outside of the widget macro.
        // This is the manual equivalent to the the menu! macro above.
        //
        // ============================================================
        //
        // relm4::menu! {
        //     main_menu: {
        //         custom: "my_widget",
        //         "Test" => TestAction,
        //         "Test2" => TestAction,
        //         "Test toggle" => TestU8Action(1_u8),
        //         section! {
        //             "Section test" => TestAction,
        //             "Test toggle" => TestU8Action(1_u8),
        //         },
        //         section! {
        //             "Test" => TestAction,
        //             "Test2" => TestAction,
        //             "Test Value" => TestU8Action(1_u8),
        //         }
        //     }
        // };

        let model = Self { counter };
        let widgets = view_output!();

        let group = RelmActionGroup::<WindowActionGroup>::new();

        let action: RelmAction<TestAction> = {
            let sender = sender.clone();
            RelmAction::new_stateless(move |_| {
                println!("Statelesss action!");
                sender.input(AppMsg::Increment);
            })
        };

        let action2: RelmAction<TestU8Action> =
            RelmAction::new_stateful_with_target_value(&0, |_, state, _value| {
                *state ^= 1;
                dbg!(state);
            });

        group.add_action(action);
        group.add_action(action2);

        let actions = group.into_action_group();
        widgets
            .main_window
            .insert_action_group("win", Some(&actions));

        ComponentParts { model, widgets }
    }

impl<Name: ActionName> RelmAction<Name> where
Name::Target: ToVariant + FromVariant,

pub fn to_menu_item_with_target_value(
label: &str,
target_value: &Name::Target
) -> MenuItem

Create a menu item for this action with the target value sent to the action on activation.

impl<Name: ActionName> RelmAction<Name> where
Name::Target: EmptyType,

pub fn to_menu_item(label: &str) -> MenuItem

Create a menu item for this action.