Module relm4::component

Components are smaller mostly independent parts of your application.

Structs

ComponentBuilder

A component that is ready for docking and launch.

ComponentParts

Contains the initial model and widgets being docked into a component.

ComponentSenderInner

Contains senders used by the component.

Connector

Contains the post-launch input sender and output receivers with the root widget.

Controller

Controls the component from afar.

StateWatcher

Keeps track of a components model and view.

Traits

Component

Elm-style variant of a Component with view updates separated from input updates

ComponentController

Shared behavior of component controller types.

OnDestroy

Type which supports signaling when it has been destroyed.

SimpleComponent

Elm-style variant of a Component with view updates separated from input updates

Type Definitions

CommandFuture

A future returned by a component’s command method.

ComponentSender

Contain senders used by the component.