Struct relm4::component::ComponentBuilder

pub struct ComponentBuilder<C: Component> {
    pub root: C::Root,
    /* private fields */
}

A component that is ready for docking and launch.

Fields
root: C::Root

The root widget of the component.

Implementations

impl<C: Component> ComponentBuilder<C>

pub fn new() -> Self

Prepares a component for initialization.

pub fn update_root<F: FnOnce(&mut C::Root)>(self, func: F) -> Self

Configure the root widget before launching.

pub fn widget(&self) -> &C::Root

Access the root widget before the component is initialized.

impl<C: Component> ComponentBuilder<C> where
C::Root: AsRef<Widget>,

pub fn attach_to(self, container: &impl RelmContainerExt) -> Self

Attach the component’s root widget to a given container.

Examples found in repository?

examples/settings_list.rs (line 26)

↕

19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69

fn main() {
    gtk::Application::builder()
        .application_id("org.relm4.SettingsListExample")
        .launch(|_app, window| {
            // Intiialize a component's root widget
            let component = App::builder()
                // Attach the root widget to the given window.
                .attach_to(&window)
                // Start the component service with an initial parameter
                .launch("Settings List Demo".into())
                // Attach the returned receiver's messages to this closure.
                .connect_receiver(move |sender, message| match message {
                    Output::Clicked(id) => {
                        eprintln!("ID {id} Clicked");

                        match id {
                            0 => xdg_open("https://github.com/AaronErhardt/Relm4".into()),
                            1 => {
                                xdg_open("https://aaronerhardt.github.io/docs/relm4/relm4/".into())
                            }
                            2 => {
                                sender.send(Input::Clear);
                            }
                            _ => (),
                        }
                    }

                    Output::Reload => {
                        sender.send(Input::AddSetting {
                            description: "Browse GitHub Repository".into(),
                            button: "GitHub".into(),
                            id: 0,
                        });

                        sender.send(Input::AddSetting {
                            description: "Browse Documentation".into(),
                            button: "Docs".into(),
                            id: 1,
                        });

                        sender.send(Input::AddSetting {
                            description: "Clear List".into(),
                            button: "Clear".into(),
                            id: 2,
                        });
                    }
                });

            println!("parent is {:?}", component.widget().toplevel_window());
        });
}

impl<C: Component> ComponentBuilder<C> where
C::Root: AsRef<Window>,

pub fn transient_for(self, window: impl AsRef<Window>) -> Self

Set the component’s root widget transient for a given window.

impl<C: Component> ComponentBuilder<C>

pub fn launch(self, payload: C::InitParams) -> Connector<C>

Starts the component, passing ownership to a future attached to a GLib context.

Examples found in repository?

examples/settings_list.rs (line 28)

↕

19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69

fn main() {
    gtk::Application::builder()
        .application_id("org.relm4.SettingsListExample")
        .launch(|_app, window| {
            // Intiialize a component's root widget
            let component = App::builder()
                // Attach the root widget to the given window.
                .attach_to(&window)
                // Start the component service with an initial parameter
                .launch("Settings List Demo".into())
                // Attach the returned receiver's messages to this closure.
                .connect_receiver(move |sender, message| match message {
                    Output::Clicked(id) => {
                        eprintln!("ID {id} Clicked");

                        match id {
                            0 => xdg_open("https://github.com/AaronErhardt/Relm4".into()),
                            1 => {
                                xdg_open("https://aaronerhardt.github.io/docs/relm4/relm4/".into())
                            }
                            2 => {
                                sender.send(Input::Clear);
                            }
                            _ => (),
                        }
                    }

                    Output::Reload => {
                        sender.send(Input::AddSetting {
                            description: "Browse GitHub Repository".into(),
                            button: "GitHub".into(),
                            id: 0,
                        });

                        sender.send(Input::AddSetting {
                            description: "Browse Documentation".into(),
                            button: "Docs".into(),
                            id: 1,
                        });

                        sender.send(Input::AddSetting {
                            description: "Clear List".into(),
                            button: "Clear".into(),
                            id: 2,
                        });
                    }
                });

            println!("parent is {:?}", component.widget().toplevel_window());
        });
}