Module relm4::factory

Defines traits and data types to generate widgets from collections efficiently.

Re-exports

pub use collections::*;

pub use positions::*;

pub use traits::*;

Modules

collections

Containers similar to std::collections that can be used to store factory data.

positions

Position types for various widgets.

traits

Traits for for managing and updating factories.

Structs

DynamicIndex

A dynamic index that updates automatically when items are shifted inside a factory container.