Module relm4::factory

Defines traits and data types used to efficiently generating widgets from collections.

Re-exports

pub use collections::*;

Modules

collections

Containers similar to std::collections that implement the Factory trait.

positions

Position types for various widgets.

Traits

Factory

A container that is a able to efficiently update, generate and remove widgets that represent the data stored in the container.

FactoryListView

Extends FactoryView for containers that work similar to lists. This means that the container can insert widgets before and after other widgets.

FactoryPrototype

Define behavior to create, update and remove widgets according to data stored in a factory.

FactoryView

A trait implemented for GTK4 widgets that allows a factory to create and remove widgets.