Module relm4::factory::collections

Containers similar to std::collections that implement the Factory trait.

Which factory type to use

Use FactoryVec if you only need to push and pop at the back. If you need more flexibility for example for pushing or removing items at arbitraty indices use FactoryVecDeque.

Structs

DynamicIndex

A dynamic index that updates automatically when items are shifted inside a Factory.

FactoryVec

A container similar to Vec that implements Factory.

FactoryVecDeque

A container similar to VecDeque that implements Factory.