Struct relm4::factory::collections::FactoryVecDeque

pub struct FactoryVecDeque<Data>where
    Data: FactoryPrototype,{ /* private fields */ }

A container similar to VecDeque that implements Factory.

Implementations

impl<Data> FactoryVecDeque<Data>where Data: FactoryPrototype,

pub fn new() -> Self

Create a new FactoryVecDeque.

pub fn from_vec_deque(data: VecDeque<Data>) -> Self

Initialize a new FactoryVecDeque with a normal VecDeque.

pub fn into_vec_deque(self) -> VecDeque<Data>

Get the internal data of the FactoryVecDeque.

pub fn clear(&mut self)

Remove all data from the FactoryVecDeque.

pub fn len(&self) -> usize

Returns the length as amount of elements stored in this type.

pub fn is_empty(&self) -> bool

Returns true if the length of this type is 0.

pub fn push_back(&mut self, data: Data)

Insert an element at the end of a FactoryVecDeque.

pub fn pop_back(&mut self) -> Option<Data>

Remove an element at the end of a FactoryVecDeque.

pub fn push_front(&mut self, data: Data)

Adds an element at the front.

pub fn pop_front(&mut self) -> Option<Data>

Removes an element from the front.

pub fn insert(&mut self, index: usize, data: Data)

Adds an element at a given index. All elements with indices greater than or equal to index will be shifted towards the back.

pub fn remove(&mut self, index: usize) -> Option<Data>

Removes an element at a given index.

pub fn get(&self, index: usize) -> Option<&Data>

Get a reference to data stored at index.

pub fn get_mut(&mut self, index: usize) -> Option<&mut Data>

Get a mutable reference to data stored at index.

Assumes that the data will be modified and the corresponding widget needs to be updated.

impl<Data, View> FactoryVecDeque<Data>where Data: FactoryPrototype<Factory = Self, View = View>, View: FactoryView<Data::Root>,

pub fn iter(&self) -> Iter<'_, Data>

Get an immutable iterator for this type