Trait relm4::factory::FactoryPrototype

pub trait FactoryPrototype: Sized {
    type Factory: Factory<Self, Self::View>;
    type Widgets;
    type Root: WidgetExt;
    type View: FactoryView<Self::Root>;
    type Msg;
    fn generate(
        &self, 
        key: &<Self::Factory as Factory<Self, Self::View>>::Key, 
        sender: Sender<Self::Msg>
    ) -> Self::Widgets;

    fn position(
        &self, 
        key: &<Self::Factory as Factory<Self, Self::View>>::Key
    ) -> <Self::View as FactoryView<Self::Root>>::Position;

    fn update(
        &self, 
        key: &<Self::Factory as Factory<Self, Self::View>>::Key, 
        widgets: &Self::Widgets
    );

    fn get_root(widgets: &Self::Widgets) -> &Self::Root;
}

Define behavior to create, update and remove widgets according to data stored in a factory.

Associated Types

type Factory: Factory<Self, Self::View>

Factory container that stores the data.

type Widgets

Type that stores all widgets needed to update them in the update function.

type Root: WidgetExt

Outermost type of the newly created widgets. Similar to the Root type in crate::Widgets.

type View: FactoryView<Self::Root>

Widget that the generated widgets are added to.

type Msg

Message type used to send messages back to the component or app where this factory is used

Required methods

fn generate(
&self,
key: &<Self::Factory as Factory<Self, Self::View>>::Key,
sender: Sender<Self::Msg>
) -> Self::Widgets

Create new widgets when self is inserted into the factory.

fn position(
&self,
key: &<Self::Factory as Factory<Self, Self::View>>::Key
) -> <Self::View as FactoryView<Self::Root>>::Position

Set the widget position upon creation, useful for gtk::Grid or similar.

fn update(
&self,
key: &<Self::Factory as Factory<Self, Self::View>>::Key,
widgets: &Self::Widgets
)

Function called when self is modified.

fn get_root(widgets: &Self::Widgets) -> &Self::Root

Get the outermost widget from the widgets.