Trait relm4::factory::FactoryPrototype

pub trait FactoryPrototype: Sized {
    type Factory: Factory<Self, Self::View>;
    type Widget: WidgetExt;
    type View: FactoryView<Self::Widget>;
    type Msg;
    fn generate(
        &self, 
        key: &<Self::Factory as Factory<Self, Self::View>>::Key, 
        sender: Sender<Self::Msg>
    ) -> Self::Widget;

    fn position(
        &self, 
        key: &<Self::Factory as Factory<Self, Self::View>>::Key
    ) -> <Self::View as FactoryView<Self::Widget>>::Position;

    fn update(
        &self, 
        key: &<Self::Factory as Factory<Self, Self::View>>::Key, 
        widget: &Self::Widget
    );

    fn remove(widget: &Self::Widget) -> &Self::Widget;
}

Define behavior to create, update and remove widgets according to data stored in a factory.

Associated Types

type Factory: Factory<Self, Self::View>

type Widget: WidgetExt

type View: FactoryView<Self::Widget>

type Msg

Required methods

fn generate(
&self,
key: &<Self::Factory as Factory<Self, Self::View>>::Key,
sender: Sender<Self::Msg>
) -> Self::Widget

Create new widgets when self is inserted into the factory.

fn position(
&self,
key: &<Self::Factory as Factory<Self, Self::View>>::Key
) -> <Self::View as FactoryView<Self::Widget>>::Position

Set the widget position upon creation, useful for gtk::Grid or similar.

fn update(
&self,
key: &<Self::Factory as Factory<Self, Self::View>>::Key,
widget: &Self::Widget
)

Function called when self is modified.

fn remove(widget: &Self::Widget) -> &Self::Widget

Function called when self is removed from the factory.