Module relm4::prelude

Commonly-imported traits and types.

Modules that contain components can glob import this module to bring all needed types and traits into scope.

Re-exports

pub use crate::factory::DynamicIndex;

pub use crate::factory::FactoryComponent;

pub use crate::factory::FactorySender;

pub use adw;

pub use gtk;

pub use panel;

Structs

ComponentParts

Contains the initial model and widgets being docked into a component.

ComponentSender

Contains senders to send and receive messages from a Component.

Controller

Controls the component from afar.

RelmApp

An app that runs the main application.

Traits

Component

The fundamental building block of a Relm4 application.

ComponentController

Shared behavior of component controller types.

RelmWidgetExt

Trait that extends gtk::prelude::WidgetExt.

SimpleComponent

Elm-style variant of a Component with view updates separated from input updates.

WidgetRef

Get a reference to a widget.

WidgetTemplate

A trait that describes a widget template.