Struct relm4::RelmApp

pub struct RelmApp<Model> where
    Model: ModelTrait + AppUpdate + 'static,
    Model::Widgets: WidgetsTrait<Model, ()> + 'static,
    <Model::Widgets as WidgetsTrait<Model, ()>>::Root: IsA<ApplicationWindow> + IsA<Window>,
    Model::Components: Components<Model> + 'static,  { /* fields omitted */ }

The app that runs the main application. A RelmApp consists of a model that stores the application state and widgets that represent the UI.

A RelmApp might run as a standalone application or may contain multiple components that communicate with each other.

Implementations

impl<Model> RelmApp<Model> where
Model: ModelTrait + AppUpdate + 'static,
Model::Widgets: WidgetsTrait<Model, ()> + 'static,
<Model::Widgets as WidgetsTrait<Model, ()>>::Root: IsA<ApplicationWindow> + IsA<Window>,
Model::Components: Components<Model> + 'static,

pub fn run(&self)

Runs the application, returns once the application is closed.

Unlike gtk::Application::run, this function does not handle command-line arguments. To pass arguments to GTK, use RelmApp::run_with_args.

Examples found in repository

relm4-examples/examples/open_button.rs (line 99)

96
97
98
99
100

fn main() {
    let model = AppModel {};
    let app = RelmApp::new(model);
    app.run();
}

relm4-examples/examples/stack_switcher.rs (line 132)

129
130
131
132
133

fn main() {
    let model = AppModel {};
    let relm = RelmApp::new(model);
    relm.run();
}

relm4-examples/libadwaita/examples/simple.rs (line 84)

81
82
83
84
85

fn main() {
    let model = AppModel::default();
    let app = RelmApp::new(model);
    app.run();
}

relm4-examples/examples/popover.rs (line 82)

79
80
81
82
83

fn main() {
    let model = AppModel::default();
    let app = RelmApp::new(model);
    app.run();
}

relm4-examples/examples/simple_override_relm4.rs (line 73)

70
71
72
73
74

fn main() {
    let model = AppModel::default();
    let app = RelmApp::new(model);
    app.run();
}

relm4-examples/examples/alert.rs (line 169)

166
167
168
169
170

fn main() {
    let model = AppModel::default();
    let app = RelmApp::new(model);
    app.run();
}

Additional examples can be found in:

relm4-examples/examples/simple.rs

relm4-examples/examples/micro_components.rs

relm4-examples/examples/save_dialog.rs

relm4-examples/examples/actions.rs

relm4-examples/examples/menu.rs

relm4-examples/examples/relm4_actions.rs

relm4-examples/examples/stateful_msg_handler.rs

relm4-examples/examples/simple_manual.rs

relm4-examples/examples/non_blocking_async.rs

relm4-examples/examples/components_old.rs

relm4-examples/examples/stack.rs

relm4-examples/examples/components.rs

relm4-examples/examples/to_do.rs

relm4-examples/libadwaita/examples/view-switcher.rs

relm4-examples/examples/grid_factory.rs

relm4-examples/examples/factory_manual.rs

relm4-examples/examples/macro.rs

relm4-examples/examples/stack_factory.rs

relm4-examples/examples/factory.rs

relm4-examples/examples/factory_advanced.rs

relm4-examples/examples/future.rs

relm4-examples/examples/drawing.rs

relm4-examples/examples/macro_test.rs

relm4-examples/examples/entry_tracker.rs

relm4-examples/examples/tracker.rs

relm4-examples/examples/tokio.rs

relm4-examples/examples/entry.rs

relm4-examples/examples/list.rs

relm4-examples/libadwaita/examples/calc-trainer.rs

relm4-examples/examples/non_blocking_async_manual.rs

relm4-examples/libadwaita/examples/advent_calendar.rs

pub fn run_with_args<S>(&self, args: &[S]) where
S: AsRef<str>,

Runs the application with the provided command-line arguments, returns once the application is closed.

pub fn new(model: Model) -> Self

Create a Relm4 application.

Examples found in repository

relm4-examples/examples/open_button.rs (line 98)

96
97
98
99
100

fn main() {
    let model = AppModel {};
    let app = RelmApp::new(model);
    app.run();
}

relm4-examples/examples/stack_switcher.rs (line 131)

129
130
131
132
133

fn main() {
    let model = AppModel {};
    let relm = RelmApp::new(model);
    relm.run();
}

relm4-examples/libadwaita/examples/simple.rs (line 83)

81
82
83
84
85

fn main() {
    let model = AppModel::default();
    let app = RelmApp::new(model);
    app.run();
}

relm4-examples/examples/popover.rs (line 81)

79
80
81
82
83

fn main() {
    let model = AppModel::default();
    let app = RelmApp::new(model);
    app.run();
}

relm4-examples/examples/simple_override_relm4.rs (line 72)

70
71
72
73
74

fn main() {
    let model = AppModel::default();
    let app = RelmApp::new(model);
    app.run();
}

relm4-examples/examples/alert.rs (line 168)

166
167
168
169
170

fn main() {
    let model = AppModel::default();
    let app = RelmApp::new(model);
    app.run();
}

Additional examples can be found in:

relm4-examples/examples/simple.rs

relm4-examples/examples/micro_components.rs

relm4-examples/examples/save_dialog.rs

relm4-examples/examples/actions.rs

relm4-examples/examples/menu.rs

relm4-examples/examples/relm4_actions.rs

relm4-examples/examples/stateful_msg_handler.rs

relm4-examples/examples/simple_manual.rs

relm4-examples/examples/non_blocking_async.rs

relm4-examples/examples/components_old.rs

relm4-examples/examples/stack.rs

relm4-examples/examples/components.rs

relm4-examples/examples/to_do.rs

relm4-examples/libadwaita/examples/view-switcher.rs

relm4-examples/examples/grid_factory.rs

relm4-examples/examples/factory_manual.rs

relm4-examples/examples/macro.rs

relm4-examples/examples/stack_factory.rs

relm4-examples/examples/factory.rs

relm4-examples/examples/factory_advanced.rs

relm4-examples/examples/future.rs

relm4-examples/examples/drawing.rs

relm4-examples/examples/macro_test.rs

relm4-examples/examples/entry_tracker.rs

relm4-examples/examples/tracker.rs

relm4-examples/examples/tokio.rs

relm4-examples/examples/entry.rs

relm4-examples/examples/list.rs

relm4-examples/libadwaita/examples/calc-trainer.rs

relm4-examples/examples/non_blocking_async_manual.rs

relm4-examples/libadwaita/examples/advent_calendar.rs

pub fn with_app(model: Model, app: Application) -> Self

Create a new Relm4 application with an existing gtk::Application.