Struct relm4::RelmApp

pub struct RelmApp { /* private fields */ }

An app that runs the main application.

Implementations

impl RelmApp

pub fn new(app_id: &str) -> Self

Create a new Relm4 application.

This function will create a new gtk::Application object if necessary.

If the libadwaita feature is enabled, then the created gtk::Application will be an instance of adw::Application. This can be overridden by passing your own application object to RelmApp::with_app.

Examples found in repository?

relm4/examples/menu.rs (line 173)

172
173
174
175

fn main() {
    let app = RelmApp::new("relm4.example.menu");
    app.run::<App>(0);
}

Hide additional examples

relm4-components/examples/alert.rs (line 146)

145
146
147
148

fn main() {
    let app = RelmApp::new("relm4.example.alert");
    app.run::<App>(());
}

relm4/examples/entry.rs (line 136)

135
136
137
138

fn main() {
    let app = RelmApp::new("relm4.example.entry");
    app.run::<App>(());
}

relm4/examples/simple.rs (line 76)

75
76
77
78

fn main() {
    let app = RelmApp::new("relm4.example.simple");
    app.run::<App>(0);
}

relm4/examples/to_do.rs (line 156)

155
156
157
158

fn main() {
    let app = RelmApp::new("relm4.example.to_do");
    app.run::<App>(());
}

relm4/examples/factory.rs (line 222)

221
222
223
224

fn main() {
    let app = RelmApp::new("relm4.example.factory");
    app.run::<App>(0);
}

Additional examples can be found in:

relm4/examples/factory_async.rs

relm4/examples/worker.rs

relm4/examples/actions.rs

relm4/examples/popover.rs

relm4/examples/drawing.rs

relm4/examples/components.rs

relm4-components/examples/open_button.rs

relm4-components/examples/web_image.rs

relm4/examples/grid_factory.rs

relm4-components/examples/file_dialogs.rs

relm4/examples/simple_manual.rs

relm4/examples/message_broker.rs

relm4/examples/message_stream.rs

relm4/examples/transient_dialog.rs

relm4/examples/progress.rs

relm4/examples/simple_async.rs

relm4/examples/non_blocking_async.rs

relm4/examples/non_blocking_sync.rs

relm4/examples/widget_template.rs

relm4/examples/macro_reference.rs

relm4/examples/tracker.rs

relm4/examples/log.rs

pub fn with_app(app: impl IsA<Application>) -> Self

Create a Relm4 application with a provided gtk::Application.

pub fn run<C>(self, payload: C::Init)where C: Component, C::Root: IsA<Window> + WidgetExt,

Runs the application, returns once the application is closed.

Unlike gtk::prelude::ApplicationExtManual::run, this function does not handle command-line arguments. To pass arguments to GTK, use RelmApp::run_with_args.

Examples found in repository?

relm4/examples/menu.rs (line 174)

172
173
174
175

fn main() {
    let app = RelmApp::new("relm4.example.menu");
    app.run::<App>(0);
}

Hide additional examples

relm4-components/examples/alert.rs (line 147)

145
146
147
148

fn main() {
    let app = RelmApp::new("relm4.example.alert");
    app.run::<App>(());
}

relm4/examples/entry.rs (line 137)

135
136
137
138

fn main() {
    let app = RelmApp::new("relm4.example.entry");
    app.run::<App>(());
}

relm4/examples/simple.rs (line 77)

75
76
77
78

fn main() {
    let app = RelmApp::new("relm4.example.simple");
    app.run::<App>(0);
}

relm4/examples/to_do.rs (line 157)

155
156
157
158

fn main() {
    let app = RelmApp::new("relm4.example.to_do");
    app.run::<App>(());
}

relm4/examples/factory.rs (line 223)

221
222
223
224

fn main() {
    let app = RelmApp::new("relm4.example.factory");
    app.run::<App>(0);
}

Additional examples can be found in:

relm4/examples/factory_async.rs

relm4/examples/worker.rs

relm4/examples/actions.rs

relm4/examples/popover.rs

relm4/examples/drawing.rs

relm4/examples/components.rs

relm4-components/examples/open_button.rs

relm4-components/examples/web_image.rs

relm4/examples/grid_factory.rs

relm4-components/examples/file_dialogs.rs

relm4/examples/simple_manual.rs

relm4/examples/message_broker.rs

relm4/examples/message_stream.rs

relm4/examples/transient_dialog.rs

relm4/examples/progress.rs

relm4/examples/non_blocking_async.rs

relm4/examples/non_blocking_sync.rs

relm4/examples/widget_template.rs

relm4/examples/macro_reference.rs

relm4/examples/tracker.rs

relm4/examples/log.rs

pub fn run_with_args<C, S>(self, payload: C::Init, args: &[S])where C: Component, C::Root: IsA<Window> + WidgetExt, S: AsRef<str>,

Runs the application with the provided command-line arguments, returns once the application is closed.

pub fn run_async<C>(self, payload: C::Init)where C: AsyncComponent, C::Root: IsA<Window> + WidgetExt,

Runs the application, returns once the application is closed.

Unlike gtk::prelude::ApplicationExtManual::run, this function does not handle command-line arguments. To pass arguments to GTK, use RelmApp::run_with_args.

Examples found in repository?

relm4/examples/simple_async.rs (line 109)

107
108
109
110

fn main() {
    let app = RelmApp::new("relm4.example.simple_async");
    app.run_async::<App>(0);
}

pub fn run_async_with_args<C, S>(self, payload: C::Init, args: &[S])where C: AsyncComponent, C::Root: IsA<Window> + WidgetExt, S: AsRef<str>,

Runs the application with the provided command-line arguments, returns once the application is closed.