Struct relm4::RelmComponent

pub struct RelmComponent<Model, ParentModel> where
    Model: ComponentUpdate<ParentModel> + 'static,
    ParentModel: ModelTrait,
    Model::Widgets: WidgetsTrait<Model, ParentModel> + 'static,  { /* fields omitted */ }

A component that can be part of the main application or other components.

A RelmComponent has its own widget, model and message type and can send messages to its parent and its children components.

Multiple RelmComponents that have the same parent are usually bundled in a struct that implements Components.

Implementations

impl<Model, ParentModel> RelmComponent<Model, ParentModel> where
Model::Widgets: WidgetsTrait<Model, ParentModel> + 'static,
ParentModel: ModelTrait,
Model: ComponentUpdate<ParentModel> + 'static,

pub fn new(
parent_model: &ParentModel,
parent_widgets: &ParentModel::Widgets,
parent_sender: Sender<ParentModel::Msg>
) -> Self

Create a new RelmComponent.

pub fn send(&self, msg: Model::Msg) -> Result<(), SendError<Model::Msg>>

Send a message to this component. This can be used by the parent to send messages to this component.

pub fn root_widget(
&self
) -> &<Model::Widgets as WidgetsTrait<Model, ParentModel>>::Root

Returns the root widget of this component’s widgets. Can be used by the parent Widgets::connect_components to connect the root widget to the parent’s widgets.

impl<Model, ParentModel> RelmComponent<Model, ParentModel> where
Model: ComponentUpdate<ParentModel> + Send + 'static,
Model::Widgets: WidgetsTrait<Model, ParentModel> + 'static,
Model::Components: Send,
Model::Msg: Send,
ParentModel: ModelTrait,
ParentModel::Msg: Send,

pub fn with_new_thread(
parent_model: &ParentModel,
parent_widgets: &ParentModel::Widgets,
parent_sender: Sender<ParentModel::Msg>
) -> Self

Create a new RelmComponent that runs the ComponentUpdate::update function in another thread.

Because GTK4 widgets are neither Send nor Sync we must still run the Widgets::view function on the main thread. Also the model needs to be sent between threads to run the update and view functions. So if you look want to send a lot of messages to self using a RelmWorker will perform better.