Trait relm4::Components

pub trait Components<ParentModel: ?Sized + Model> {
    fn init_components(
        parent_model: &ParentModel, 
        parent_widget: &ParentModel::Widgets, 
        parent_sender: Sender<ParentModel::Msg>
    ) -> Self;
}

Define how to initialize one or more components.

Typically a struct is used to store multiple components that are child components of the app or another component.

Example

struct AppComponents {
    comp1: RelmComponent<Comp1Model, AppModel>,
    comp2: RelmComponent<Comp2Model, AppModel>,
}

impl Components<AppModel> for AppComponents {
    fn init_components(parent_model: &AppModel, parent_widgets: &AppWidgets, parent_sender: Sender<AppMsg>) -> Self {
        AppComponents {
            comp1: RelmComponent::with_new_thread(parent_model, parent_widgets, parent_sender.clone()),
            comp2: RelmComponent::new(parent_model, parent_widgets, parent_sender),
        }
    }
}

Required methods

fn init_components(
parent_model: &ParentModel,
parent_widget: &ParentModel::Widgets,
parent_sender: Sender<ParentModel::Msg>
) -> Self