Trait relm4::ContainerChild

pub trait ContainerChild {
    type Child: IsA<Widget>;
}

Type of children inside a container.

For example, gtk::ListBox only contains gtk::ListBoxRow widgets as children. If you add any other kind of widget, a row is automatically inserted between the list box and the widget.

For simple widgets like gtk::Box, the children type will be gtk::Widget, meaning that it can be any widget type.

Required Associated Types

type Child: IsA<Widget>

Type of container children.