Trait relm4::RelmRemoveExt

pub trait RelmRemoveExt: ContainerChild {
    fn container_remove(&self, widget: &impl AsRef<Self::Child>);
}

Widget types which can have widgets removed from them.

Required Methods

fn container_remove(&self, widget: &impl AsRef<Self::Child>)

Removes the widget from the container if it is a child of the container.