Trait relm4::RelmSetChildExt

pub trait RelmSetChildExt: ContainerChild {
    fn container_set_child(&self, widget: Option<&impl AsRef<Widget>>);
}

Widget types which allow to set or unset their child.

Required Methods

fn container_set_child(&self, widget: Option<&impl AsRef<Widget>>)

Set a child for the container or remove it using None.