Trait relm4::RelmSetChildExt

pub trait RelmSetChildExt: ContainerChild {
    // Required methods
    fn container_set_child(&self, widget: Option<&impl AsRef<Widget>>);
    fn container_get_child(&self) -> Option<Widget>;
}

Widget types which allow to set or unset their child.

Required Methods

fn container_set_child(&self, widget: Option<&impl AsRef<Widget>>)

Set a child for the container or remove it using None.

fn container_get_child(&self) -> Option<Widget>

Get the child of a container. Returns None if the container has no child.